Conservation problem class

This class is used to represent a data instance for the multi-action conservation planning problem. This includes several methods for obtaining instance information. It is created using the problem function.

Fields

$data

list object containing data.

Methods

$getActionsAmount( )

return integer number of possible actions.

$getData(character name)

return an data.frame object stored in the data field with the corresponding name. The argument name indicates the name of arguments of the problem function ("pu", "features", "rij", "threats", "sensibility" or "bound").

$getFeatureAmount( )

return integer number of features.

$getFeatureNames( )

return character names of features.

$getPlanningUnitCosts( )

return the numeric vector cost of allocating each planning unit.

$getPlanningUnitsAmount( )

return integer number of planning units.

$getThreatCosts( )

return the numeric vector cost of actions each planning unit and threat.

$getThreatNames( )

return character names of threats.

$getThreatsAmount( )

return integer number of threats.

$print( )

print basic information of the data instance.

$show( )

call print method.

Examples

## Examples of how to use the methods of a ConservationProblem class object.

## Load data
data(example_pu_data, example_features_data, example_rij_data, example_threats_data, example_sensibility_data, example_bound_data)

## Create data instance
problem_data <- problem(
  pu = example_pu_data, features = example_features_data, rij = example_rij_data,
  threats = example_threats_data, sensibility = example_sensibility_data,
  bound = example_bound_data
)
#> Warning: The following pu's do not contain species: 59 97 109 118 121 122 123 127 129 134 136 139 141 142 146 166 167 185 194 195 196 198 211 230
#> Correctly loaded inputs

## Use class methods
problem_data$getActionsAmount()
#> [1] 562

problem_data$getData("features")
#>    id target   name internal_id
#> 1   1    100 scl_ja           1
#> 2   2    100 nem_er           2
#> 3   3     83 thr_sc           3
#> 4   4     51 neo_be           4
#> 5   5     29 neo_gr           5
#> 6   6     28 neo_le           6
#> 7   7    100 neo_pa           7
#> 8   8    100 ano_da           8
#> 9   9    100 neo_at           9
#> 10 10     54 neo_hy          10

problem_data$getFeatureAmount()
#> [1] 10

problem_data$getFeatureNames()
#>  [1] "scl_ja" "nem_er" "thr_sc" "neo_be" "neo_gr" "neo_le" "neo_pa" "ano_da"
#>  [9] "neo_at" "neo_hy"

problem_data$getPlanningUnitCosts()
#>   [1] 3 4 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3
#>  [38] 3 3 3 3 3 3 3 3 3 3 3 3 3 3 4 3 3 3 3 3 3 5 3 3 3 5 3 3 3 3 3 3 3 3 3 3 3
#>  [75] 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 4 4 3 3 3 3 3 3 5 5 3 3 3 5 3 3
#> [112] 3 3 3 3 3 3 5 3 3 5 5 5 3 3 4 3 3 5 3 3 3 3 5 3 5 3 3 3 3 5 4 3 3 3 4 3 3
#> [149] 3 3 4 3 4 4 3 3 3 3 3 3 3 4 4 3 3 4 4 3 4 4 4 3 4 4 3 4 4 3 4 3 3 3 4 3 5
#> [186] 4 3 4 3 4 4 4 3 5 3 5 3 5 4 3 3 3 4 3 3 3 3 3 3 4 5 5 4 3 4 4 3 3 3 4 3 3
#> [223] 3 4 3 3 4 3 4 4 4 4 3 3 4 3 3 3 3 5

problem_data$getPlanningUnitsAmount()
#> [1] 240

problem_data$getThreatCosts()
#>   [1] 3 3 4 4 4 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3
#>  [38] 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3
#>  [75] 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 4 4 4 3 3 3 3 3
#> [112] 3 3 3 3 3 3 3 5 5 5 5 3 3 3 3 3 3 5 5 5 5 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3
#> [149] 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3
#> [186] 3 3 3 3 3 3 3 3 3 3 3 4 4 4 4 4 4 3 3 3 3 3 3 3 3 3 3 3 3 5 5 5 5 5 5 5 5
#> [223] 3 3 3 3 3 3 5 5 5 5 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 5 5 5 5 3 3 3 3 5 5 5
#> [260] 5 5 5 5 5 5 5 5 5 3 3 3 3 4 4 4 3 3 3 3 5 5 5 5 3 3 3 3 3 3 3 3 5 5 5 5 3
#> [297] 3 5 5 5 5 3 3 3 3 3 3 3 3 5 5 5 5 4 4 4 3 3 3 3 3 3 4 4 4 3 3 3 3 3 3 3 3
#> [334] 4 4 4 3 3 4 4 4 4 4 4 3 3 3 3 3 3 3 3 3 3 3 3 3 3 4 4 4 4 4 4 3 3 3 3 4 4
#> [371] 4 4 4 4 3 3 4 4 4 4 4 4 4 4 4 3 3 4 4 4 4 4 4 3 3 4 4 4 4 4 4 3 3 4 4 4 3
#> [408] 3 3 3 3 3 4 4 4 3 3 5 5 5 5 4 4 4 3 3 4 4 4 3 3 4 4 4 4 4 4 4 4 4 3 3 5 5
#> [445] 5 5 3 3 5 5 5 5 3 3 5 5 5 5 4 4 4 3 3 3 3 3 3 4 4 4 3 3 3 3 3 3 3 3 3 3 3
#> [482] 3 4 4 4 5 5 5 5 5 5 5 5 4 4 4 3 3 4 4 4 4 4 4 3 3 3 3 3 3 4 4 4 3 3 3 3 3
#> [519] 3 4 4 4 3 3 3 3 4 4 4 3 3 4 4 4 4 4 4 4 4 4 4 4 4 3 3 3 3 4 4 4 3 3 3 3 3
#> [556] 3 3 3 5 5 5 5

problem_data$getThreatNames()
#> [1] "1" "2" "3" "4"

problem_data$getThreatsAmount()
#> [1] 4

problem_data$print()
#> Conservation Problem
#>   planning units: data.frame (240 units)
#>   unit costs:     min: 3, max: 5
#>   features:       scl_ja, nem_er, thr_sc, ... (10 features)
#>   threats:        1, 2, 3, 4 (4 threats)
#>   threat costs:   min: 3, max: 5