Optimization problem class

This class is used to represent an optimization model. This includes several methods for obtaining model information. It is created used models functions (e.g. min costs).

Fields

$data

list object containing data of the mathematical model.

$ConservationClass

object of class ConservationProblem-class that contain the data instance.

Methods

$getData(character name)

return an vector object stored in the data field with the corresponding name. The data correspond to the different parts of the mathematical model. The argument name can be made to the following: "obj", "rhs", "sense", "vtype", "A", "bounds" or "modelsense".

$getDataList( )

returns an list of vector object stored in the data. It correspond to set of data stored asociated to the mathematical model.

$getModelSense( )

returns a character indicating whether the model is minimization or maximization.

$getNcol( )

returns an integer number indicating the columns of matrix A.

$getNrow( )

returns an integer number indicating the rows of matrix A.

$getSizeA( )

returns a double number indicating the size of matrix A (in Megabytes).

$print( )

print basic information of the optimization model.

$show( )

call print method.

Examples

## Examples of how to use the methods of a OptimizationProblem class object.

## Load data
data(example_pu_data, example_features_data, example_rij_data, example_threats_data, example_sensibility_data, example_bound_data)

## Create data instance
problem_data <- problem(
  pu = example_pu_data, features = example_features_data, rij = example_rij_data,
  threats = example_threats_data, sensibility = example_sensibility_data,
  bound = example_bound_data
)
#> Warning: The following pu's do not contain species: 59 97 109 118 121 122 123 127 129 134 136 139 141 142 146 166 167 185 194 195 196 198 211 230
#> Correctly loaded inputs

## Create optimization model
problem_model <- min_costs(problem_data, blm = 1, blm_actions = 1)

## Use class methods
head(problem_model$getData("obj"))
#> [1] 3.000000 6.284457 5.284457 5.284457 5.284457 7.284457

head(problem_model$getDataList()[7])
#> $modelsense
#> [1] "min"
#> 

problem_model$getModelSense()
#> [1] "minimization"

problem_model$getNcol()
#> [1] 11740

problem_model$getNrow()
#> [1] 18004

problem_model$getSizeA()
#> 0.7 bytes

problem_model$print()
#> Optimization Problem
#>   model sense: minimization
#>   dimensions:  18004, 11740, 0.658 Mb (nrow, ncol, size)
#>   variables:   11740
#>   building time: 0.05 sec