Conservation problem class

This class is used to represent a data instance for the multi-action conservation planning problem. This includes several methods for obtaining instance information. It is created using the problem() function.

Fields

$data

list object containing data.

Methods

$getActionsAmount( )

return integer number of possible actions.

$getData(character name)

return an data.frame() object stored in the data field with the corresponding name. The argument name indicates the name of arguments of the problem function ("pu", "features", "rij", "threats", "sensitivity" or "boundary").

$getFeatureAmount( )

return integer number of features.

$getFeatureNames( )

return character names of features.

$getPlanningUnitCosts( )

return the numeric vector() cost of allocating each planning unit.

$getPlanningUnitsAmount( )

return integer number of planning units.

$getThreatCosts( )

return the numeric vector() cost of actions each planning unit and threat.

$getThreatNames( )

return character names of threats.

$getThreatsAmount( )

return integer number of threats.

$print( )

print basic information of the data instance.

$show( )

call print method.

Examples

## Examples of how to use the methods of a ConservationProblem class object.

## Load data
data(example_pu_data, example_features_data, example_rij_data, example_threats_data, example_sensitivity_data, example_boundary_data)
#> Warning: data set ‘example_rij_data’ not found
#> Warning: data set ‘example_boundary_data’ not found

## Create data instance
problem_data <- problem(
  pu = example_pu_data, features = example_features_data, dist_features = example_dist_features__data,
  threats = example_threats_data, dist_therats = example_dist_threats_data, sensitivity = example_sensibility_data,
  boundary = example_boundary_data
)
#> Error in h(simpleError(msg, call)) error in evaluating the argument 'dist_features' in selecting a method for function 'problem': object 'example_dist_features__data' not found

## Use class methods
problem_data$getData("features")
#> Error in eval(expr, envir, enclos) object 'problem_data' not found

problem_data$getFeatureAmount()
#> Error in eval(expr, envir, enclos) object 'problem_data' not found

problem_data$getFeatureNames()
#> Error in eval(expr, envir, enclos) object 'problem_data' not found

problem_data$getPlanningUnitCosts()
#> Error in eval(expr, envir, enclos) object 'problem_data' not found

problem_data$getPlanningUnitsAmount()
#> Error in eval(expr, envir, enclos) object 'problem_data' not found

problem_data$getThreatCosts()
#> Error in eval(expr, envir, enclos) object 'problem_data' not found

problem_data$getThreatNames()
#> Error in eval(expr, envir, enclos) object 'problem_data' not found

problem_data$getThreatsAmount()
#> Error in eval(expr, envir, enclos) object 'problem_data' not found

problem_data$print()
#> Error in eval(expr, envir, enclos) object 'problem_data' not found