Evaluate multiple blm values

Provides multiple solutions for different values of blm. Like prioriactions() function, it inherits all arguments from problem(), model() and solve().

evalBlm(data = list(), name_model = "minimizeCosts", values = c(), ...)

Arguments

data
list. Input data list for problem() function.

name_model
character. Name of the type of model to create. With two possible values: "minimizeCosts" and "maximizeBenefits".

values
numeric. Values of blm to verify. More than one value is needed.

...
arguments inherited from problem(), minimizeCosts(), maximizeBenefits(), and solve() functions.

Value

An object of class portfolio.

Details

evalblm() creates and solves multiple instances, of the corresponding multi-actions planning problem, for different values of blm. Alternatively, this could be obtained by executing function prioriactions() or by steps the problem(), model() and solve() functions; using, in each run, different blm values. However, the evalblm() function has two advantages with respect to this manual approach: : 1) it is more efficient to create the models (this is because the model is created just once and, at each iteration, only the blm values are updated); and 2) the output is a portfolio object, which allows obtaining information about the group of solutions (including all get functions).

Examples

# set seed for reproducibility
set.seed(14)

## Load data
inputs <- list(sim_pu_data, sim_features_data, sim_dist_features_data,
sim_threats_data, sim_dist_threats_data, sim_sensitivity_data,
sim_boundary_data)

## Create model and solve
port <- evalBlm(data = inputs, values = c(0.0, 0.5, 4, 8),
                name_model = "minimizeCosts", time_limit = 50, output_file = FALSE)
#> *********************************
#>  Iteration 1 of 4: Blm0
#> *********************************
#> Warning: The blm argument was set to 0, so the boundary data has no effect
#> Warning: Some blm_actions argument were set to 0, so the boundary data has no effect for these cases
#> Gurobi Optimizer version 9.1.2 build v9.1.2rc0 (linux64)
#> Thread count: 2 physical cores, 4 logical processors, using up to 2 threads
#> Optimize a model with 280 rows, 589 columns and 822 nonzeros
#> Model fingerprint: 0x9475cbf3
#> Model has 176 general constraints
#> Variable types: 352 continuous, 237 integer (237 binary)
#> Coefficient statistics:
#>   Matrix range     [5e-01, 2e+00]
#>   Objective range  [1e+00, 1e+01]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [8e+00, 2e+01]
#> Found heuristic solution: objective 1002.0000000
#> Presolve added 289 rows and 635 columns
#> Presolve time: 0.02s
#> Presolved: 569 rows, 1224 columns, 2970 nonzeros
#> Presolved model has 176 SOS constraint(s)
#> Variable types: 1056 continuous, 168 integer (162 binary)
#> 
#> Root relaxation: objective 1.455000e+02, 186 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0  145.50000    0   12 1002.00000  145.50000  85.5%     -    0s
#> H    0     0                     973.0000000  145.50000  85.0%     -    0s
#>      0     0  148.00000    0    6  973.00000  148.00000  84.8%     -    0s
#> H    0     0                     162.0000000  148.00000  8.64%     -    0s
#>      0     0  149.00000    0    6  162.00000  149.00000  8.02%     -    0s
#> H    0     0                     152.0000000  149.00000  1.97%     -    0s
#>      0     0  149.50000    0    5  152.00000  149.50000  1.64%     -    0s
#> H    0     0                     150.0000000  149.50000  0.33%     -    0s
#> 
#> Explored 1 nodes (219 simplex iterations) in 0.07 seconds
#> Thread count was 2 (of 4 available processors)
#> 
#> Solution count 5: 150 152 162 ... 1002
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective 1.500000000000e+02, best bound 1.500000000000e+02, gap 0.0000%
#> *********************************
#>  Iteration 2 of 4: Blm0.5
#> *********************************
#> Warning: Some blm_actions argument were set to 0, so the boundary data has no effect for these cases
#> Gurobi Optimizer version 9.1.2 build v9.1.2rc0 (linux64)
#> Thread count: 2 physical cores, 4 logical processors, using up to 2 threads
#> Optimize a model with 29980 rows, 10489 columns and 70122 nonzeros
#> Model fingerprint: 0x6ad01d4d
#> Model has 176 general constraints
#> Variable types: 352 continuous, 10137 integer (10137 binary)
#> Coefficient statistics:
#>   Matrix range     [5e-01, 2e+00]
#>   Objective range  [5e-01, 7e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 2e+01]
#> Found heuristic solution: objective 26936.361106
#> Presolve added 352 rows and 698 columns
#> Presolve time: 0.17s
#> Presolved: 30332 rows, 11187 columns, 72396 nonzeros
#> Presolved model has 176 SOS constraint(s)
#> Variable types: 1056 continuous, 10131 integer (10125 binary)
#> 
#> Root simplex log...
#> 
#> Iteration    Objective       Primal Inf.    Dual Inf.      Time
#>     6452    8.8063758e+03   1.825000e+02   0.000000e+00      1s
#>     7620    8.8719008e+03   0.000000e+00   0.000000e+00      1s
#> 
#> Root relaxation: objective 8.871901e+03, 7620 iterations, 1.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 8871.90080    0 4613 26936.3611 8871.90080  67.1%     -    1s
#> H    0     0                    9596.3338377 8871.90080  7.55%     -    1s
#> H    0     0                    8993.1009457 8871.90080  1.35%     -    1s
#>      0     0 8876.24862    0 4603 8993.10095 8876.24862  1.30%     -    2s
#>      0     0 8876.24862    0 2083 8993.10095 8876.24862  1.30%     -    3s
#>      0     0 8883.34676    0 2043 8993.10095 8883.34676  1.22%     -    4s
#>      0     0 8897.99101    0 1951 8993.10095 8897.99101  1.06%     -    4s
#>      0     0 8900.09761    0 1945 8993.10095 8900.09761  1.03%     -    4s
#>      0     0 8918.47875    0 1726 8993.10095 8918.47875  0.83%     -    6s
#>      0     0 8918.47875    0 2018 8993.10095 8918.47875  0.83%     -    6s
#>      0     0 8918.47875    0 1941 8993.10095 8918.47875  0.83%     -    7s
#>      0     0 8918.47875    0 1715 8993.10095 8918.47875  0.83%     -    7s
#>      0     0 8918.77414    0 1665 8993.10095 8918.77414  0.83%     -    7s
#>      0     0 8919.29314    0 1657 8993.10095 8919.29314  0.82%     -    7s
#>      0     0 8919.29369    0 1658 8993.10095 8919.29369  0.82%     -    7s
#>      0     0 8927.87544    0 1906 8993.10095 8927.87544  0.73%     -    8s
#>      0     0 8927.96410    0 1839 8993.10095 8927.96410  0.72%     -    9s
#>      0     0 8928.01350    0 1852 8993.10095 8928.01350  0.72%     -    9s
#>      0     0 8928.02673    0 1857 8993.10095 8928.02673  0.72%     -    9s
#>      0     0 8930.66787    0 1848 8993.10095 8930.66787  0.69%     -    9s
#>      0     0 8931.28621    0 1793 8993.10095 8931.28621  0.69%     -    9s
#>      0     0 8931.42513    0 1852 8993.10095 8931.42513  0.69%     -   10s
#>      0     0 8931.52078    0 1851 8993.10095 8931.52078  0.68%     -   10s
#>      0     0 8931.55095    0 1852 8993.10095 8931.55095  0.68%     -   10s
#>      0     0 8931.56782    0 1852 8993.10095 8931.56782  0.68%     -   10s
#>      0     0 8932.74330    0 1746 8993.10095 8932.74330  0.67%     -   10s
#>      0     0 8932.85512    0 1807 8993.10095 8932.85512  0.67%     -   10s
#>      0     0 8932.89035    0 1866 8993.10095 8932.89035  0.67%     -   10s
#>      0     0 8932.89843    0 1805 8993.10095 8932.89843  0.67%     -   10s
#>      0     0 8933.08371    0 1464 8993.10095 8933.08371  0.67%     -   11s
#>      0     0 8933.19078    0 1639 8993.10095 8933.19078  0.67%     -   11s
#>      0     0 8933.22263    0 1519 8993.10095 8933.22263  0.67%     -   11s
#>      0     0 8933.25311    0 1526 8993.10095 8933.25311  0.67%     -   11s
#>      0     0 8933.26237    0 1525 8993.10095 8933.26237  0.67%     -   11s
#>      0     0 8933.30659    0 1770 8993.10095 8933.30659  0.66%     -   11s
#>      0     0 8933.31827    0 1571 8993.10095 8933.31827  0.66%     -   11s
#>      0     0 8933.33040    0 1717 8993.10095 8933.33040  0.66%     -   11s
#>      0     0 8933.33040    0 1717 8993.10095 8933.33040  0.66%     -   11s
#> H    0     0                    8990.6897887 8933.33040  0.64%     -   12s
#>      0     0 8933.33040    0 1925 8990.68979 8933.33040  0.64%     -   12s
#>      0     0 8933.33040    0 1827 8990.68979 8933.33040  0.64%     -   13s
#>      0     0 8933.33040    0 1729 8990.68979 8933.33040  0.64%     -   13s
#>      0     0 8933.33040    0 1533 8990.68979 8933.33040  0.64%     -   13s
#>      0     0 8934.07813    0 1609 8990.68979 8934.07813  0.63%     -   14s
#>      0     0 8934.31218    0 1615 8990.68979 8934.31218  0.63%     -   14s
#>      0     0 8934.36032    0 1563 8990.68979 8934.36032  0.63%     -   14s
#>      0     0 8934.37293    0 1564 8990.68979 8934.37293  0.63%     -   14s
#>      0     0 8934.37712    0 1613 8990.68979 8934.37712  0.63%     -   14s
#>      0     0 8934.37791    0 1614 8990.68979 8934.37791  0.63%     -   14s
#>      0     0 8935.39998    0 1473 8990.68979 8935.39998  0.61%     -   14s
#>      0     0 8935.42901    0 1477 8990.68979 8935.42901  0.61%     -   14s
#>      0     0 8936.01441    0 1480 8990.68979 8936.01441  0.61%     -   14s
#>      0     0 8936.09961    0 1477 8990.68979 8936.09961  0.61%     -   14s
#>      0     0 8936.10620    0 1523 8990.68979 8936.10620  0.61%     -   15s
#>      0     0 8936.68999    0 1866 8990.68979 8936.68999  0.60%     -   15s
#>      0     0 8936.92280    0 1867 8990.68979 8936.92280  0.60%     -   15s
#>      0     0 8936.95113    0 1867 8990.68979 8936.95113  0.60%     -   15s
#>      0     0 8937.45863    0 1527 8990.68979 8937.45863  0.59%     -   15s
#>      0     0 8937.52826    0 1483 8990.68979 8937.52826  0.59%     -   16s
#>      0     0 8937.55477    0 1531 8990.68979 8937.55477  0.59%     -   16s
#>      0     0 8937.71841    0 1686 8990.68979 8937.71841  0.59%     -   16s
#>      0     0 8937.81409    0 1373 8990.68979 8937.81409  0.59%     -   16s
#>      0     0 8937.83307    0 1701 8990.68979 8937.83307  0.59%     -   16s
#>      0     0 8938.28668    0 1692 8990.68979 8938.28668  0.58%     -   16s
#>      0     0 8938.43400    0 1693 8990.68979 8938.43400  0.58%     -   16s
#>      0     0 8938.56552    0 1738 8990.68979 8938.56552  0.58%     -   17s
#>      0     0 8938.60863    0 1752 8990.68979 8938.60863  0.58%     -   17s
#>      0     0 8938.64284    0 1750 8990.68979 8938.64284  0.58%     -   17s
#>      0     0 8938.67307    0 1743 8990.68979 8938.67307  0.58%     -   17s
#>      0     0 8938.85214    0 1742 8990.68979 8938.85214  0.58%     -   17s
#>      0     0 8938.91414    0 1746 8990.68979 8938.91414  0.58%     -   17s
#>      0     0 8938.94906    0 1739 8990.68979 8938.94906  0.58%     -   17s
#>      0     0 8938.95926    0 1689 8990.68979 8938.95926  0.58%     -   17s
#>      0     0 8938.98846    0 1689 8990.68979 8938.98846  0.58%     -   17s
#>      0     0 8939.00295    0 1695 8990.68979 8939.00295  0.57%     -   17s
#>      0     0 8939.01377    0 1687 8990.68979 8939.01377  0.57%     -   17s
#>      0     0 8939.01377    0 1686 8990.68979 8939.01377  0.57%     -   18s
#>      0     2 8939.01377    0 1686 8990.68979 8939.01377  0.57%     -   18s
#>      3     6 8964.52518    2 1695 8990.68979 8940.60274  0.56%  1868   19s
#>      7     6 8970.61244    3 1518 8990.68979 8948.35513  0.47%  1572   20s
#>     17     6     cutoff    5      8990.68979 8956.62718  0.38%  1122   21s
#>     33     8     cutoff   11      8990.68979 8973.66671  0.19%   758   22s
#> 
#> Cutting planes:
#>   Gomory: 6
#>   Mod-K: 2
#>   RLT: 872
#>   BQP: 564
#> 
#> Explored 131 nodes (85102 simplex iterations) in 22.95 seconds
#> Thread count was 2 (of 4 available processors)
#> 
#> Solution count 4: 8990.69 8993.1 9596.33 26936.4 
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective 8.990689788719e+03, best bound 8.990689788719e+03, gap 0.0000%
#> *********************************
#>  Iteration 3 of 4: Blm4
#> *********************************
#> Warning: Some blm_actions argument were set to 0, so the boundary data has no effect for these cases
#> Gurobi Optimizer version 9.1.2 build v9.1.2rc0 (linux64)
#> Thread count: 2 physical cores, 4 logical processors, using up to 2 threads
#> Optimize a model with 29980 rows, 10489 columns and 70122 nonzeros
#> Model fingerprint: 0x2ee1f0a6
#> Model has 176 general constraints
#> Variable types: 352 continuous, 10137 integer (10137 binary)
#> Coefficient statistics:
#>   Matrix range     [5e-01, 2e+00]
#>   Objective range  [1e+00, 6e+03]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 2e+01]
#> Found heuristic solution: objective 208476.88885
#> Presolve added 352 rows and 698 columns
#> Presolve time: 0.22s
#> Presolved: 30332 rows, 11187 columns, 72396 nonzeros
#> Presolved model has 176 SOS constraint(s)
#> Variable types: 1056 continuous, 10131 integer (10125 binary)
#> 
#> Root simplex log...
#> 
#> Iteration    Objective       Primal Inf.    Dual Inf.      Time
#>     4399    6.6010047e+04   2.701000e+03   0.000000e+00      1s
#>     8125    6.9763382e+04   0.000000e+00   0.000000e+00      2s
#> 
#> Root relaxation: objective 6.976338e+04, 8125 iterations, 1.38 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 69763.3817    0 4612 208476.889 69763.3817  66.5%     -    2s
#> H    0     0                    75405.670701 69763.3817  7.48%     -    2s
#> H    0     0                    70712.807565 69763.3817  1.34%     -    2s
#>      0     0 69796.5743    0 4602 70712.8076 69796.5743  1.30%     -    3s
#>      0     0 69796.5743    0 2138 70712.8076 69796.5743  1.30%     -    4s
#>      0     0 69865.3078    0 2098 70712.8076 69865.3078  1.20%     -    4s
#>      0     0 69971.2258    0 1895 70712.8076 69971.2258  1.05%     -    5s
#>      0     0 70066.2905    0 1645 70712.8076 70066.2905  0.91%     -    5s
#>      0     0 70122.3250    0 2005 70712.8076 70122.3250  0.84%     -    6s
#>      0     0 70122.3250    0 1925 70712.8076 70122.3250  0.84%     -    7s
#>      0     0 70122.3250    0 1853 70712.8076 70122.3250  0.84%     -    8s
#>      0     0 70122.3250    0 1778 70712.8076 70122.3250  0.84%     -    8s
#>      0     0 70122.3250    0 1653 70712.8076 70122.3250  0.84%     -    8s
#>      0     0 70128.4890    0 1720 70712.8076 70128.4890  0.83%     -    9s
#>      0     0 70128.5924    0 1731 70712.8076 70128.5924  0.83%     -    9s
#>      0     0 70172.4422    0 1865 70712.8076 70172.4422  0.76%     -   10s
#>      0     0 70172.4422    0 1925 70712.8076 70172.4422  0.76%     -   10s
#>      0     0 70172.4422    0 1845 70712.8076 70172.4422  0.76%     -   11s
#>      0     0 70172.4422    0 1709 70712.8076 70172.4422  0.76%     -   11s
#>      0     0 70172.4422    0 1752 70712.8076 70172.4422  0.76%     -   11s
#>      0     0 70175.7971    0 1841 70712.8076 70175.7971  0.76%     -   11s
#>      0     0 70177.2408    0 1848 70712.8076 70177.2408  0.76%     -   11s
#>      0     0 70177.3458    0 1851 70712.8076 70177.3458  0.76%     -   11s
#>      0     0 70177.3918    0 1850 70712.8076 70177.3918  0.76%     -   11s
#>      0     0 70198.7765    0 1477 70712.8076 70198.7765  0.73%     -   12s
#>      0     0 70198.9411    0 1484 70712.8076 70198.9411  0.73%     -   12s
#>      0     0 70212.5165    0 1888 70712.8076 70212.5165  0.71%     -   13s
#>      0     0 70213.0046    0 1886 70712.8076 70213.0046  0.71%     -   13s
#>      0     0 70213.0709    0 1882 70712.8076 70213.0709  0.71%     -   13s
#>      0     0 70218.6140    0 1725 70712.8076 70218.6140  0.70%     -   13s
#>      0     0 70218.9040    0 1723 70712.8076 70218.9040  0.70%     -   13s
#>      0     0 70225.3292    0 1839 70712.8076 70225.3292  0.69%     -   13s
#>      0     0 70226.4644    0 1893 70712.8076 70226.4644  0.69%     -   14s
#>      0     0 70226.9761    0 1898 70712.8076 70226.9761  0.69%     -   14s
#>      0     0 70227.5385    0 1888 70712.8076 70227.5385  0.69%     -   14s
#>      0     0 70227.6475    0 1898 70712.8076 70227.6475  0.69%     -   14s
#>      0     0 70229.9126    0 1898 70712.8076 70229.9126  0.68%     -   14s
#>      0     0 70230.3406    0 1897 70712.8076 70230.3406  0.68%     -   14s
#>      0     0 70230.4640    0 1893 70712.8076 70230.4640  0.68%     -   14s
#>      0     0 70231.1035    0 1903 70712.8076 70231.1035  0.68%     -   14s
#>      0     0 70231.3679    0 1894 70712.8076 70231.3679  0.68%     -   15s
#>      0     0 70231.6666    0 1890 70712.8076 70231.6666  0.68%     -   15s
#>      0     0 70231.7201    0 1892 70712.8076 70231.7201  0.68%     -   15s
#>      0     0 70231.9996    0 1894 70712.8076 70231.9996  0.68%     -   15s
#>      0     0 70231.9996    0 1894 70712.8076 70231.9996  0.68%     -   15s
#>      0     2 70231.9996    0 1894 70712.8076 70231.9996  0.68%     -   15s
#>      1     5 70301.9468    1 1657 70712.8076 70232.0025  0.68%  1837   16s
#>     17     9 70565.0956    5 1303 70712.8076 70410.5450  0.43%   649   17s
#>     66    13 70599.2142    4 1580 70712.8076 70483.0668  0.32%   318   18s
#> 
#> Cutting planes:
#>   Gomory: 5
#>   Zero half: 1
#>   Mod-K: 1
#>   RLT: 594
#>   BQP: 166
#> 
#> Explored 109 nodes (70351 simplex iterations) in 18.58 seconds
#> Thread count was 2 (of 4 available processors)
#> 
#> Solution count 3: 70712.8 75405.7 208477 
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective 7.071280756525e+04, best bound 7.071280756525e+04, gap 0.0000%
#> *********************************
#>  Iteration 4 of 4: Blm8
#> *********************************
#> Warning: Some blm_actions argument were set to 0, so the boundary data has no effect for these cases
#> Gurobi Optimizer version 9.1.2 build v9.1.2rc0 (linux64)
#> Thread count: 2 physical cores, 4 logical processors, using up to 2 threads
#> Optimize a model with 29980 rows, 10489 columns and 70122 nonzeros
#> Model fingerprint: 0x316b1735
#> Model has 176 general constraints
#> Variable types: 352 continuous, 10137 integer (10137 binary)
#> Coefficient statistics:
#>   Matrix range     [5e-01, 2e+00]
#>   Objective range  [1e+00, 1e+04]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 2e+01]
#> Found heuristic solution: objective 415951.77770
#> Presolve added 352 rows and 698 columns
#> Presolve time: 0.16s
#> Presolved: 30332 rows, 11187 columns, 72396 nonzeros
#> Presolved model has 176 SOS constraint(s)
#> Variable types: 1056 continuous, 10131 integer (10125 binary)
#> 
#> Root simplex log...
#> 
#> Iteration    Objective       Primal Inf.    Dual Inf.      Time
#>     6797    1.3860285e+05   1.730000e+02   0.000000e+00      1s
#>     7912    1.3935340e+05   0.000000e+00   0.000000e+00      1s
#> 
#> Root relaxation: objective 1.393534e+05, 7912 iterations, 1.03 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 139353.395    0 4613 415951.778 139353.395  66.5%     -    1s
#> H    0     0                    150616.34140 139353.395  7.48%     -    1s
#> H    0     0                    141249.61513 139353.395  1.34%     -    1s
#>      0     0 139407.257    0 4602 141249.615 139407.257  1.30%     -    2s
#>      0     0 139407.257    0 2138 141249.615 139407.257  1.30%     -    3s
#>      0     0 139540.827    0 2098 141249.615 139540.827  1.21%     -    4s
#>      0     0 139771.430    0 1938 141249.615 139771.430  1.05%     -    4s
#>      0     0 140008.637    0 1792 141249.615 140008.637  0.88%     -    5s
#>      0     0 140008.962    0 1793 141249.615 140008.962  0.88%     -    5s
#>      0     0 140201.846    0 2088 141249.615 140201.846  0.74%     -    7s
#>      0     0 140201.846    0 1901 141249.615 140201.846  0.74%     -    7s
#>      0     0 140201.846    0 1819 141249.615 140201.846  0.74%     -    8s
#>      0     0 140201.846    0 1672 141249.615 140201.846  0.74%     -    8s
#>      0     0 140201.846    0 1801 141249.615 140201.846  0.74%     -    8s
#>      0     0 140214.820    0 1771 141249.615 140214.820  0.73%     -    9s
#>      0     0 140219.490    0 1830 141249.615 140219.490  0.73%     -    9s
#>      0     0 140220.503    0 1829 141249.615 140220.503  0.73%     -    9s
#>      0     0 140220.558    0 1829 141249.615 140220.558  0.73%     -    9s
#>      0     0 140277.177    0 1784 141249.615 140277.177  0.69%     -    9s
#>      0     0 140279.453    0 1785 141249.615 140279.453  0.69%     -   10s
#>      0     0 140280.306    0 1785 141249.615 140280.306  0.69%     -   10s
#>      0     0 140280.768    0 1786 141249.615 140280.768  0.69%     -   10s
#>      0     0 140280.853    0 1784 141249.615 140280.853  0.69%     -   10s
#>      0     0 140299.166    0 1775 141249.615 140299.166  0.67%     -   10s
#>      0     0 140302.619    0 1783 141249.615 140302.619  0.67%     -   10s
#>      0     0 140303.690    0 1800 141249.615 140303.690  0.67%     -   10s
#>      0     0 140303.989    0 1797 141249.615 140303.989  0.67%     -   10s
#>      0     0 140314.550    0 1777 141249.615 140314.550  0.66%     -   11s
#>      0     0 140316.691    0 1768 141249.615 140316.691  0.66%     -   11s
#>      0     0 140317.805    0 1783 141249.615 140317.805  0.66%     -   11s
#>      0     0 140318.196    0 1778 141249.615 140318.196  0.66%     -   11s
#>      0     0 140322.352    0 1786 141249.615 140322.352  0.66%     -   11s
#>      0     0 140323.403    0 1782 141249.615 140323.403  0.66%     -   11s
#>      0     0 140324.216    0 1788 141249.615 140324.216  0.66%     -   12s
#>      0     0 140324.489    0 1788 141249.615 140324.489  0.65%     -   12s
#>      0     0 140325.430    0 1788 141249.615 140325.430  0.65%     -   12s
#>      0     0 140325.694    0 1790 141249.615 140325.694  0.65%     -   12s
#>      0     0 140326.129    0 1779 141249.615 140326.129  0.65%     -   12s
#>      0     0 140326.129    0 1779 141249.615 140326.129  0.65%     -   12s
#>      0     0 140326.129    0 1888 141249.615 140326.129  0.65%     -   13s
#>      0     0 140326.129    0 1778 141249.615 140326.129  0.65%     -   13s
#>      0     0 140326.129    0 1541 141249.615 140326.129  0.65%     -   14s
#>      0     0 140326.129    0 1684 141249.615 140326.129  0.65%     -   14s
#>      0     0 140326.129    0 1699 141249.615 140326.129  0.65%     -   14s
#>      0     0 140334.286    0 1510 141249.615 140334.286  0.65%     -   14s
#>      0     0 140337.235    0 1689 141249.615 140337.235  0.65%     -   14s
#>      0     0 140338.379    0 1567 141249.615 140338.379  0.65%     -   14s
#>      0     0 140338.567    0 1701 141249.615 140338.567  0.64%     -   14s
#>      0     0 140338.593    0 1703 141249.615 140338.593  0.64%     -   14s
#>      0     0 140347.955    0 1808 141249.615 140347.955  0.64%     -   15s
#>      0     0 140348.519    0 1693 141249.615 140348.519  0.64%     -   15s
#>      0     0 140348.663    0 1754 141249.615 140348.663  0.64%     -   15s
#>      0     0 140356.921    0 1652 141249.615 140356.921  0.63%     -   15s
#>      0     0 140358.255    0 1646 141249.615 140358.255  0.63%     -   15s
#>      0     0 140358.824    0 1652 141249.615 140358.824  0.63%     -   15s
#>      0     0 140358.972    0 1654 141249.615 140358.972  0.63%     -   16s
#>      0     0 140364.623    0 1707 141249.615 140364.623  0.63%     -   16s
#>      0     0 140365.497    0 1707 141249.615 140365.497  0.63%     -   16s
#>      0     0 140365.666    0 1703 141249.615 140365.666  0.63%     -   16s
#>      0     0 140370.165    0 1703 141249.615 140370.165  0.62%     -   16s
#>      0     0 140371.531    0 1699 141249.615 140371.531  0.62%     -   16s
#>      0     0 140371.809    0 1702 141249.615 140371.809  0.62%     -   16s
#>      0     0 140375.299    0 1702 141249.615 140375.299  0.62%     -   16s
#>      0     0 140376.566    0 1807 141249.615 140376.566  0.62%     -   17s
#>      0     0 140377.481    0 1796 141249.615 140377.481  0.62%     -   17s
#>      0     0 140377.937    0 1759 141249.615 140377.937  0.62%     -   17s
#>      0     0 140380.249    0 1620 141249.615 140380.249  0.62%     -   17s
#>      0     0 140380.893    0 1663 141249.615 140380.893  0.62%     -   17s
#>      0     0 140381.096    0 1659 141249.615 140381.096  0.61%     -   17s
#>      0     0 140381.551    0 1748 141249.615 140381.551  0.61%     -   17s
#>      0     0 140381.551    0 1748 141249.615 140381.551  0.61%     -   17s
#>      0     2 140381.551    0 1748 141249.615 140381.551  0.61%     -   18s
#>      3     6 140755.306    2 1588 141249.615 140409.679  0.59%  1851   19s
#>      5     8 140839.411    2 1602 141249.615 140409.713  0.59%  2343   20s
#>      7    10 140770.796    3 1612 141249.615 140565.409  0.48%  2369   21s
#>     11    13     cutoff    4      141249.615 140565.437  0.48%  1895   22s
#>     14    14     cutoff    5      141249.615 140565.437  0.48%  1738   23s
#>     21    13     cutoff    7      141249.615 140646.715  0.43%  1458   24s
#>     26    17 140990.522    5 1524 141249.615 140646.715  0.43%  1571   25s
#>     35    18     cutoff    6      141249.615 140646.715  0.43%  1301   26s
#>     39    14     cutoff    7      141249.615 140647.410  0.43%  1254   27s
#>     51    14     cutoff    4      141249.615 140647.410  0.43%  1161   28s
#>     57     8     cutoff    5      141249.615 140714.023  0.38%  1148   29s
#>     67     4     cutoff    5      141249.615 140839.561  0.29%  1077   30s
#>     75     4 141126.710    4 1632 141249.615 141010.488  0.17%  1012   31s
#> 
#> Cutting planes:
#>   Gomory: 6
#>   Implied bound: 1
#>   MIR: 5
#>   Flow cover: 2
#>   GUB cover: 1
#>   Mod-K: 1
#>   RLT: 694
#>   BQP: 466
#> 
#> Explored 87 nodes (133066 simplex iterations) in 31.79 seconds
#> Thread count was 2 (of 4 available processors)
#> 
#> Solution count 3: 141250 150616 415952 
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective 1.412496151305e+05, best bound 1.412496151305e+05, gap 0.0000%