Evaluate multiple target values

Provides multiple solutions for different targets values. This function assumes that the minimizeCosts model is being used. As well as the prioriactions() function, it inherits all arguments from problem(), model() and solve().

evalTarget(data = list(), values = c(), ...)

Arguments

data
list. Input data list for problem() function.

values
numeric. Proportion of maximum value of benefit to verify. This information can be getting with getBenefit() function. More than one value is needed.

...
arguments inherited from problem(), minimizeCosts(), and solve() functions.

Value

An object of class portfolio.

Details

evalTarget() creates and solves multiple instances, of the corresponding multi-actions planning problem, for different proportions of maximum benefit values as target values. Alternatively, this could be obtained by executing function prioriactions() or by steps the problem(), model() and solve() functions; using, in each run, different budgtes values. However, the evalTarget() function has two advantages with respect to this manual approach: : 1) it is more efficient to create the models (this is because the model is created just once and, at each iteration, only the target values are updated); and 2) the output is a portfolio object, which allows obtaining information about the group of solutions (including all get functions).

Examples

# set seed for reproducibility
set.seed(14)

## Load data
inputs <- list(sim_pu_data, sim_features_data, sim_dist_features_data,
sim_threats_data, sim_dist_threats_data, sim_sensitivity_data,
sim_boundary_data)

## Create model and solve
port <- evalTarget(data = inputs, values = c(0.1, 0.3, 0.5), time_limit = 50, output_file = FALSE)
#> *********************************
#>  Iteration 1 of 3: Prop0.1
#> *********************************
#> Warning: The blm argument was set to 0, so the boundary data has no effect
#> Warning: Some blm_actions argument were set to 0, so the boundary data has no effect for these cases
#> Gurobi Optimizer version 9.1.2 build v9.1.2rc0 (linux64)
#> Thread count: 2 physical cores, 4 logical processors, using up to 2 threads
#> Optimize a model with 280 rows, 589 columns and 822 nonzeros
#> Model fingerprint: 0x3e7128e8
#> Model has 176 general constraints
#> Variable types: 352 continuous, 237 integer (237 binary)
#> Coefficient statistics:
#>   Matrix range     [5e-01, 2e+00]
#>   Objective range  [1e+00, 1e+01]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [3e+00, 7e+00]
#> Found heuristic solution: objective 1002.0000000
#> Presolve added 289 rows and 635 columns
#> Presolve time: 0.00s
#> Presolved: 569 rows, 1224 columns, 2970 nonzeros
#> Presolved model has 176 SOS constraint(s)
#> Variable types: 1056 continuous, 168 integer (162 binary)
#> 
#> Root relaxation: objective 5.230000e+01, 143 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0   52.30000    0   25 1002.00000   52.30000  94.8%     -    0s
#> H    0     0                     973.0000000   52.30000  94.6%     -    0s
#>      0     0   55.50000    0   20  973.00000   55.50000  94.3%     -    0s
#> H    0     0                      80.0000000   55.50000  30.6%     -    0s
#>      0     0   55.90000    0   11   80.00000   55.90000  30.1%     -    0s
#> H    0     0                      77.0000000   55.90000  27.4%     -    0s
#>      0     0   56.20000    0   14   77.00000   56.20000  27.0%     -    0s
#> H    0     0                      68.0000000   56.20000  17.4%     -    0s
#> H    0     0                      63.0000000   56.20000  10.8%     -    0s
#>      0     0   56.60000    0   20   63.00000   56.60000  10.2%     -    0s
#>      0     0   56.60000    0   20   63.00000   56.60000  10.2%     -    0s
#>      0     0   56.60000    0   15   63.00000   56.60000  10.2%     -    0s
#>      0     0   56.90000    0   23   63.00000   56.90000  9.68%     -    0s
#>      0     0   57.20000    0   23   63.00000   57.20000  9.21%     -    0s
#> H    0     0                      62.0000000   57.20000  7.74%     -    0s
#>      0     0   57.20000    0   23   62.00000   57.20000  7.74%     -    0s
#>      0     0   57.24444    0   20   62.00000   57.24444  7.67%     -    0s
#>      0     0   57.24444    0   11   62.00000   57.24444  7.67%     -    0s
#>      0     0   57.24444    0   22   62.00000   57.24444  7.67%     -    0s
#>      0     0   57.24444    0   25   62.00000   57.24444  7.67%     -    0s
#>      0     0   57.46780    0   33   62.00000   57.46780  7.31%     -    0s
#>      0     0   57.48439    0   25   62.00000   57.48439  7.28%     -    0s
#>      0     0   57.48439    0   33   62.00000   57.48439  7.28%     -    0s
#>      0     0   57.48439    0   27   62.00000   57.48439  7.28%     -    0s
#>      0     0   57.48439    0   27   62.00000   57.48439  7.28%     -    0s
#>      0     2   57.48439    0   27   62.00000   57.48439  7.28%     -    0s
#> 
#> Cutting planes:
#>   Gomory: 4
#>   Implied bound: 1
#>   MIR: 5
#>   Flow cover: 6
#>   Inf proof: 1
#>   Relax-and-lift: 1
#> 
#> Explored 428 nodes (2081 simplex iterations) in 0.34 seconds
#> Thread count was 2 (of 4 available processors)
#> 
#> Solution count 7: 62 63 68 ... 1002
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective 6.200000000000e+01, best bound 6.200000000000e+01, gap 0.0000%
#> *********************************
#>  Iteration 2 of 3: Prop0.3
#> *********************************
#> Gurobi Optimizer version 9.1.2 build v9.1.2rc0 (linux64)
#> Thread count: 2 physical cores, 4 logical processors, using up to 2 threads
#> Optimize a model with 280 rows, 589 columns and 822 nonzeros
#> Model fingerprint: 0x9ac23276
#> Model has 176 general constraints
#> Variable types: 352 continuous, 237 integer (237 binary)
#> Coefficient statistics:
#>   Matrix range     [5e-01, 2e+00]
#>   Objective range  [1e+00, 1e+01]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [9e+00, 2e+01]
#> Found heuristic solution: objective 1002.0000000
#> Presolve added 289 rows and 635 columns
#> Presolve time: 0.00s
#> Presolved: 569 rows, 1224 columns, 2970 nonzeros
#> Presolved model has 176 SOS constraint(s)
#> Variable types: 1056 continuous, 168 integer (162 binary)
#> 
#> Root relaxation: objective 1.896000e+02, 224 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0  189.60000    0   40 1002.00000  189.60000  81.1%     -    0s
#> H    0     0                     957.0000000  189.60000  80.2%     -    0s
#>      0     0  198.52882    0   16  957.00000  198.52882  79.3%     -    0s
#> H    0     0                     208.0000000  198.52882  4.55%     -    0s
#>      0     0  198.52882    0   20  208.00000  198.52882  4.55%     -    0s
#> H    0     0                     205.0000000  198.52882  3.16%     -    0s
#>      0     0  198.52882    0   23  205.00000  198.52882  3.16%     -    0s
#>      0     0  198.57686    0   25  205.00000  198.57686  3.13%     -    0s
#>      0     0  198.57778    0   20  205.00000  198.57778  3.13%     -    0s
#>      0     0  198.66000    0   20  205.00000  198.66000  3.09%     -    0s
#>      0     0  198.75860    0   23  205.00000  198.75860  3.04%     -    0s
#>      0     0  198.80699    0   24  205.00000  198.80699  3.02%     -    0s
#>      0     0  198.89488    0   27  205.00000  198.89488  2.98%     -    0s
#>      0     0  198.93333    0   17  205.00000  198.93333  2.96%     -    0s
#>      0     0  198.93333    0   17  205.00000  198.93333  2.96%     -    0s
#>      0     0  199.16000    0   15  205.00000  199.16000  2.85%     -    0s
#>      0     0  199.16000    0   20  205.00000  199.16000  2.85%     -    0s
#>      0     0  199.20370    0   20  205.00000  199.20370  2.83%     -    0s
#>      0     0  199.20370    0   20  205.00000  199.20370  2.83%     -    0s
#>      0     0  199.20370    0   25  205.00000  199.20370  2.83%     -    0s
#>      0     0  199.20370    0   18  205.00000  199.20370  2.83%     -    0s
#>      0     2  199.20370    0   18  205.00000  199.20370  2.83%     -    0s
#>   2223   648  202.63333   24   19  205.00000  201.64034  1.64%   3.8    1s
#>   4536   570  202.33750   29   22  205.00000  201.77222  1.57%   3.9    2s
#>   8382   218  203.60000   36    7  205.00000  203.60000  0.68%   3.9    3s
#> 
#> Cutting planes:
#>   Gomory: 4
#>   Implied bound: 1
#>   MIR: 6
#>   Flow cover: 25
#>   Inf proof: 9
#> 
#> Explored 10318 nodes (39779 simplex iterations) in 3.32 seconds
#> Thread count was 2 (of 4 available processors)
#> 
#> Solution count 4: 205 208 957 1002 
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Warning: max constraint violation (2.8000e-02) exceeds tolerance
#> Warning: max general constraint violation (2.8000e-02) exceeds tolerance
#>          (model may be infeasible or unbounded - try turning presolve off)
#> Best objective 2.050000000000e+02, best bound 2.050000000000e+02, gap 0.0000%
#> *********************************
#>  Iteration 3 of 3: Prop0.5
#> *********************************
#> Gurobi Optimizer version 9.1.2 build v9.1.2rc0 (linux64)
#> Thread count: 2 physical cores, 4 logical processors, using up to 2 threads
#> Optimize a model with 280 rows, 589 columns and 822 nonzeros
#> Model fingerprint: 0x1766737a
#> Model has 176 general constraints
#> Variable types: 352 continuous, 237 integer (237 binary)
#> Coefficient statistics:
#>   Matrix range     [5e-01, 2e+00]
#>   Objective range  [1e+00, 1e+01]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [2e+01, 3e+01]
#> Found heuristic solution: objective 1002.0000000
#> Presolve added 289 rows and 635 columns
#> Presolve time: 0.00s
#> Presolved: 569 rows, 1224 columns, 2970 nonzeros
#> Presolved model has 176 SOS constraint(s)
#> Variable types: 1056 continuous, 168 integer (162 binary)
#> 
#> Root relaxation: objective 3.605000e+02, 201 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0  360.50000    0   30 1002.00000  360.50000  64.0%     -    0s
#> H    0     0                     918.0000000  360.50000  60.7%     -    0s
#>      0     0  370.00000    0    6  918.00000  370.00000  59.7%     -    0s
#> H    0     0                     384.0000000  370.00000  3.65%     -    0s
#>      0     0  370.00000    0    4  384.00000  370.00000  3.65%     -    0s
#> H    0     0                     377.0000000  370.00000  1.86%     -    0s
#>      0     0  370.00000    0    3  377.00000  370.00000  1.86%     -    0s
#>      0     0  370.00000    0    3  377.00000  370.00000  1.86%     -    0s
#>      0     0  370.00000    0    3  377.00000  370.00000  1.86%     -    0s
#>      0     2  370.00000    0    3  377.00000  370.00000  1.86%     -    0s
#>   2287   796     cutoff   31       377.00000  371.00000  1.59%   2.3    1s
#>   6869  2370  375.69388   27   13  377.00000  372.00000  1.33%   2.5    2s
#>  10754  3675  374.00000   24    4  377.00000  372.00000  1.33%   2.6    3s
#>  15184  4713  374.50000   30   17  377.00000  372.63910  1.16%   2.7    4s
#>  19588  5149  375.00000   26    3  377.00000  373.00000  1.06%   2.8    5s
#>  23525  5756  376.00000   34   14  377.00000  373.00000  1.06%   2.9    6s
#>  27192  6244     cutoff   33       377.00000  373.00000  1.06%   2.9    7s
#>  32090  6944  374.00000   27    3  377.00000  373.50000  0.93%   2.9    8s
#>  35124  7144     cutoff   29       377.00000  374.00000  0.80%   2.9    9s
#>  38977  6629  374.00000   24   11  377.00000  374.00000  0.80%   3.0   10s
#>  42076  6462  375.50000   30   10  377.00000  374.00000  0.80%   3.1   11s
#>  45541  6483     cutoff   32       377.00000  374.00000  0.80%   3.1   12s
#>  49754  6516     cutoff   28       377.00000  374.25000  0.73%   3.1   13s
#>  53494  6023  376.00000   30    4  377.00000  374.50000  0.66%   3.2   14s
#>  57361  4662     cutoff   25       377.00000  375.00000  0.53%   3.2   15s
#>  61100  3876  376.00000   31    9  377.00000  375.00000  0.53%   3.2   16s
#>  64285  3298     cutoff   27       377.00000  375.00000  0.53%   3.3   17s
#>  66032  2396  376.00000   34    3  377.00000  375.50000  0.40%   3.3   18s
#>  68781   868     cutoff   26       377.00000  376.00000  0.27%   3.3   19s
#> 
#> Cutting planes:
#>   Gomory: 2
#>   Flow cover: 4
#>   Inf proof: 23
#> 
#> Explored 70348 nodes (234405 simplex iterations) in 19.35 seconds
#> Thread count was 2 (of 4 available processors)
#> 
#> Solution count 4: 377 384 918 1002 
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective 3.770000000000e+02, best bound 3.770000000000e+02, gap 0.0000%