Extract connectivity values

Provides the connectivity value for all actions and planning units in a solution.

getConnectivity(x)

Arguments

x
solution or portfolio object.

Value

data.frame.

Details

The connectivity among is calculated as the sum of all connectivity penalties by each action and planning unit in the solution. This can be expressed mathematically for a set of planning units \(I\) indexed by \(i\) and \(j\), and a set of threats \(K\) indexed by \(k\) as:

$$ \sum_{k \in K}\sum_{i \in I_k}\sum_{j \in I_k} x_{ik} (1 - x_{jk})cv_{ij} $$

Where, \(x_{ik}\) is the decisions variable that specify whether an action has been selected to abate threat \(k\) in planning unit \(i\) (1) or not (0), \(cv_{ij}\) is the connectivity penalty that applies when a solution contains planning unit \(i\) but not \(j\) o viceversa.

Note that there is an action per threat, so it is assumed that the index of the threat coincides with the index of the action used to abate it.

Examples

# set seed for reproducibility
set.seed(14)

## Load data
data(sim_pu_data, sim_features_data, sim_dist_features_data,
sim_threats_data, sim_dist_threats_data, sim_sensitivity_data,
sim_boundary_data)

## Create data instance
problem_data <- inputData(
  pu = sim_pu_data, features = sim_features_data, dist_features = sim_dist_features_data,
  threats = sim_threats_data, dist_threats = sim_dist_threats_data,
  sensitivity = sim_sensitivity_data, boundary = sim_boundary_data
)

## Create optimization model
problem_model <- problem(x = problem_data, blm = 0.03)
#> Warning: Some blm_actions argument were set to 0, so the boundary data has no effect for these cases

## Solve the optimization model
s <- solve(a = problem_model, solver = "gurobi", time_limit = 20, output_file = FALSE)
#> Gurobi Optimizer version 9.1.2 build v9.1.2rc0 (linux64)
#> Thread count: 2 physical cores, 4 logical processors, using up to 2 threads
#> Optimize a model with 30160 rows, 10472 columns and 70437 nonzeros
#> Model fingerprint: 0xde6e27dc
#> Variable types: 352 continuous, 10120 integer (10120 binary)
#> Coefficient statistics:
#>   Matrix range     [5e-01, 2e+00]
#>   Objective range  [3e-02, 3e+01]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 2e+01]
#> Presolve removed 366 rows and 357 columns
#> Presolve time: 0.18s
#> Presolved: 29794 rows, 10115 columns, 69709 nonzeros
#> Variable types: 0 continuous, 10115 integer (10110 binary)
#> Found heuristic solution: objective 717.5310881
#> Found heuristic solution: objective 427.0047919
#> 
#> Root relaxation: objective 3.581240e+02, 8746 iterations, 3.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0  358.12398    0 10000  427.00479  358.12398  16.1%     -    3s
#> H    0     0                     420.5230312  358.12398  14.8%     -    3s
#>      0     0  359.91688    0 9986  420.52303  359.91688  14.4%     -    5s
#>      0     0  359.91688    0 5059  420.52303  359.91688  14.4%     -    7s
#> H    0     0                     416.2575327  359.91688  13.5%     -    7s
#>      0     0  360.10196    0 5058  416.25753  360.10196  13.5%     -    8s
#> H    0     0                     412.3820503  360.10196  12.7%     -    8s
#>      0     0  360.28513    0 5067  412.38205  360.28513  12.6%     -    8s
#>      0     0  360.39906    0 5068  412.38205  360.39906  12.6%     -    9s
#>      0     0  360.43049    0 5060  412.38205  360.43049  12.6%     -   10s
#>      0     0  360.50949    0 5060  412.38205  360.50949  12.6%     -   10s
#>      0     0  360.50967    0 5068  412.38205  360.50967  12.6%     -   10s
#>      0     0  360.62212    0 5067  412.38205  360.62212  12.6%     -   10s
#> H    0     0                     412.2729579  360.62212  12.5%     -   10s
#>      0     0  360.66568    0 5068  412.27296  360.66568  12.5%     -   11s
#> H    0     0                     410.9323026  360.66568  12.2%     -   11s
#>      0     0  360.69499    0 5060  410.93230  360.69499  12.2%     -   11s
#>      0     0  360.76829    0 5068  410.93230  360.76829  12.2%     -   11s
#>      0     0  360.85420    0 5067  410.93230  360.85420  12.2%     -   12s
#>      0     0  360.86860    0 5068  410.93230  360.86860  12.2%     -   12s
#>      0     0  360.92688    0 5066  410.93230  360.92688  12.2%     -   12s
#> H    0     0                     409.1440845  360.92688  11.8%     -   12s
#>      0     0  360.97794    0 5067  409.14408  360.97794  11.8%     -   12s
#>      0     0  361.01607    0 5067  409.14408  361.01607  11.8%     -   13s
#> H    0     0                     408.7142319  361.01607  11.7%     -   13s
#>      0     0  361.10380    0 5070  408.71423  361.10380  11.6%     -   13s
#>      0     0  361.13613    0 5072  408.71423  361.13613  11.6%     -   14s
#>      0     0  361.14546    0 5072  408.71423  361.14546  11.6%     -   14s
#>      0     0  361.14546    0 5072  408.71423  361.14546  11.6%     -   14s
#> H    0     0                     407.2598499  361.14546  11.3%     -   15s
#> H    0     0                     407.1802042  361.14546  11.3%     -   15s
#>      0     2  361.14546    0 5072  407.18020  361.14546  11.3%     -   15s
#> 
#> Cutting planes:
#>   Gomory: 3
#>   Zero half: 3
#>   RLT: 4150
#>   BQP: 63
#> 
#> Explored 5 nodes (20814 simplex iterations) in 20.00 seconds
#> Thread count was 2 (of 4 available processors)
#> 
#> Solution count 10: 407.18 407.26 408.714 ... 427.005
#> 
#> Time limit reached
#> Best objective 4.071802041812e+02, best bound 3.646037157994e+02, gap 10.4564%

# get connectivities values
getConnectivity(s)
#> Error in getConnectivity(s) could not find function "getConnectivity"