Optimization problem class

This class encodes the corresponding optimization model. It is created using `model()`` function.

Fields

$data

list object containing data of the mathematical model.

$ConservationClass

object of class conservationProblem-class() that contains the data structure.

Methods

getData(character name)

vector(). Object stored in the data field with the corresponding name. The data correspond to the different parts of the mathematical model. The argument name can be made to the following: "obj", "rhs", "sense", "vtype", "A", "bounds" or "modelsense".

getDataList()

list() of vector(). Object stored in the data. It contains all information relative to the mathematical model, such as "obj", "rhs", etc.getModelSense() character. Indicate whether the model is minimization or maximization.getNcol() integer. Number indicating the columns of matrix A.getNrow() integer. Number indicating the rows of matrix A.getSizeA() character. Number indicating the size of matrix A (in kilo Bytes).print() Print basic information of the optimization model.show() Call print method.

# set seed for reproducibility set.seed(14)## Load data data(sim_pu_data, sim_features_data, sim_dist_features_data, sim_threats_data, sim_dist_threats_data, sim_sensitivity_data, sim_boundary_data)## Create data instance problem_data <- problem( pu = sim_pu_data, features = sim_features_data, dist_features = sim_dist_features_data, threats = sim_threats_data, dist_threats = sim_dist_threats_data, sensitivity = sim_sensitivity_data, boundary = sim_boundary_data )## Create optimization model problem_model <- minimizeCosts(x = problem_data, blm = 1)## Use class methods head(problem_model$getData("obj"))head(problem_model$getDataList()[7])problem_model$getModelSense()problem_model$getNconstraints()problem_model$getNvariables()problem_model$getSizeA()problem_model$print()