The prioriactions R package

The prioriactions R package uses a mixed integer mathematical programming (MIP) approach for building and solving multi-action conservation planning problems, where the goal is to find an optimal combination of management actions that abate threats, in an efficient way while accounting for connectivity. Furthermore, the flexibility of the package interface allows implementing an extended version of the base model for minimizing fragmentation between different actions. These models were called as MAMP and MAMP-E by Salgado-Rojas et al. (2020), where you can get a detailed description of how both problems were mathematically modeled.

Once built a multi-action conservation problem, the next step is to solve it. For this, the package has a variety of commercial and open-source exact algorithm solvers that guarantee to find optimal solutions. The models implemented here have three advantages over their heuristic counterparts: shorter execution times, higher solutions quality, and a solution quality guarantee.

Create an optimization model for the multi-action conservation planning problem, following the mathematical formulations used in Salgado-Rojas et al. (2020). This function is used to specify model configuration parameters related to connectivity issues and its internal modus operandi.

Connectivity parameters (blm and blm_actions) manipulate the spatial fragmentation of planning units and/or action management to improve the compactness of the reserve solutions. Likewise, the other parameters (curve and segments) affect the linearization strategy used by the model internally in order to be solved under a linear programming approach. It is not recommended to modify the default values of the later ones.

prioriactions(data = list(), name_model = "minimizeCosts", ...)

prioriactions(data = list(), name_model = "minimizeCosts", ...)

Arguments

x
Object of class ConservationProblem-class() that specifies the basic data used in a problem of prioritization of multiple conservation actions. This object must be created using the problem() function.

blm
A numeric value that indicates the penalty factor associated to the spatial fragmentation of planning units, similar to Boundary Length Modifier (BLM) in Marxan. This argument only has an effect when the bound() argument of problem() function is a data.frame object. The default argument is zero.

blm_actions
A numeric value that indicates the penalty factor associated to the spatial fragmentation of actions. The default argument is zero.

curve
An integer value that selects the type of continuous curve that will represent the expression (linear or non-linear) associated with a specific constraint in this model. Therefore, the curve can represent a linear (1), quadratic (2) or cubic (3) function. The default argument is 3 and it is not recommended to change this value unless you have advanced knowledge of the linearization of mathematical model.

segments
An integer value that selects the number of segments (1, 2 or 3) that will have the piecewise linear function in charge of approximating the non-linear expression contained in a specific constraint of this model. The default argument is 3 and it is not recommended to change this value unless you have advanced knowledge of the linearization of mathematical model.

Value

An object of class OptimizationProblem-class().

Details

Put details here!

Put details here! The details may include the mathematical formulation of the optimization model associated with this conservation problem and/or a rough description of the mathematical model, and/or what happens when the parameters are set in one way or another.

References

Salgado-Rojas J, <U+00C1>lvarez-Miranda E, Hermoso V, Garcia-Gonzalo J, Weintraub A. A mixed integer programming approach for multi-action planning for threat management. Ecological Modelling 2020; 418:108901.
(DOI: https://doi.org/10.1016/j.ecolmodel.2019.108901)

See also

Examples

## Create an optimization model for the multi-action conservation
## planning problem using a data instance that has been created in R.
## This example uses input files included into package.

## Load package
library(prioriactions)

## Load data
data(example_pu_data, example_features_data, example_dist_features_data, example_dist_threats_data, example_threats_data, example_sensitivity_data, example_bound_data)
#> Warning: data set ‘example_pu_data’ not found
#> Warning: data set ‘example_features_data’ not found
#> Warning: data set ‘example_dist_features_data’ not found
#> Warning: data set ‘example_dist_threats_data’ not found
#> Warning: data set ‘example_threats_data’ not found
#> Warning: data set ‘example_sensitivity_data’ not found
#> Warning: data set ‘example_bound_data’ not found

## Create data instance
input <- list(
  pu = example_pu_data, features = example_features_data, dist_features = example_dist_features_data,
  threats = example_threats_data, dist_threats = example_dist_threats_data, sensitivity = example_sensitivity_data,
  bound = example_bound_data
)
#> Error in eval(expr, envir, enclos) object 'example_pu_data' not found

## Create optimization model
portfolio <- prioriactions(data = input, name_model = "minimizeCosts", blm = 0, output_file = FALSE, time_limit = 10)
#> Error in eval(assertion, env) object 'input' not found