Multi-action conservation planning problem

Create a data instance for the multi-action conservation planning problem. This function is used to specify the basic data used in a problem of prioritization of multiple conservation actions. This includes, the spatial distribution of the planning units and their costs, the features to be conserved (e.g. species, ecosystems) and their threats, and the costs of implementing an action to remove such threats.

problem(pu, features, dist_features, threats, dist_threats, sensitivity, ...)

# S4 method for data.frame
problem(pu, features, dist_features, threats, dist_threats, sensitivity, boundary = NULL)

Arguments

pu
Object of class data.frame() that specifies the planning units to use in the reserve design exercise and their corresponding cost. It may be desirable to exclude some planning units from the analysis, for example those outside the study area. To exclude planning units you have two options: (i) pre-process the data.frame before using it (i.e., manually exclude planning units); or (ii) set the status of planning units to 3 not to include them. Each row corresponds to a different planning unit, and each column corresponds to different information about the planning units. The description of the columns is listed below.

"id"

integer unique identifier for each planning unit. This column is mandatory.

"cost"

numeric cost of including each planning unit in the reserve system. This column is mandatory.

"status"

integer indicating if each planning unit should not be locked in the solution (0) or if it should be locked-in (2) or locked-out (3) of the solution. This column is optional.

features
Object of class data.frame() that specifies the conservation features to consider in the reserve design exercise, such as their name and representation requirements (i.e., targets). Each row corresponds to a different feature and each column corresponds to different information about the features. The description of the columns is listed below.

"id"

integer unique identifier for each conservation feature. This column is mandatory.

"target"

numeric target amount for each conservation feature. This column is mandatory.

"name"

character name for each conservation feature. This column is optional.

dist_features
Object of class data.frame() that specifies the spatial distribution of conservation features across planning units. It is recommended not to include cases where a feature does not occur in a planning unit (avoid rows with "amount" equal to 0). Instead, it is recommended to omit these cases completely from the data.frame(), for efficiency reasons. The description of the columns is listed below.

"pu"

integer id of a planning unit where the conservation feature listed on the same row occurs. This column is mandatory.

"species"

integer id of each conservation feature. This column is mandatory.

"amount"

integer indicating if the conservation feature occurring in the planning unit listed on the same row (1), or if it is not occurring (0). This column is mandatory.

threats
Object of class data.frame() that specifies the threats to consider in the reserve design exercise, such as their name and penalty of connectivity (i.e., blm_actions). Each row corresponds to a different feature and each column corresponds to different information about the threats The description of the columns is listed below.

"id"

integer unique identifier for each threat. This column is mandatory.

"name"

character name for each threat. This column is optional.

"blm_actions"

numeric penalty of connectivity between threats. This column is optional.

dist_threats
Object of class data.frame() that specifies the spatial distribution of threats across planning units and the corresponding cost of applying an action against these threats. It is recommended not to include cases where a threat does not occur in a planning unit (avoid rows with "amount" equal to 0). Instead, it is recommended to omit these cases completely from the data.frame(), for efficiency reasons. The description of the columns is listed below.

"pu"

integer id of a planning unit where the threat listed on the same row occurs. This column is mandatory.

"threats"

integer id of each threat. This column is mandatory.

"amount"

integer indicating if the threat occurring in the planning unit listed on the same row (1), or if it is not occurring (0). This column is mandatory.

"cost"

numeric cost of applying an action to eliminate the threat in the planning unit listed in the same row. This column is mandatory.

"status"

integer indicating if each action should not be locked in the solution (0) or if it should be locked-in (2) or locked-out (3) of the solution. This column is optional.

sensitivity
Object of class data.frame() that specifies the species-threats sensitivity, i.e., which threats affect the persistence of endangered conservation features (e.g., species). It is recommended not to include cases where there is no species-threats sensitivity (avoid rows with "amount" equal to 0). Instead, it is recommended to omit these cases completely from the data.frame(), for efficiency reasons. The description of the columns is listed below.

"species"

integer id of each conservation feature. This column is mandatory.

"threats"

integer id of each threat. This column is mandatory.

"amount"

integer indicating if the threat againsting the conservation feature listed on the same row (1), or if it is not againsting (0). This column is mandatory.

boundary
Object of class NULL() (default) or an object of class data.frame(). NULL indicates that planning units' boundaries (i.e., boundary data) are not required for the multi-action conservation planning problem. data.frame specifies the boundaries of the planning units, i.e., the spatial relationship between two units, such as the "length" of the shared boundary between them. The description of the columns is listed below.

"id1"

integer id of each planning unit. This column is mandatory.

"id2"

integer id of each planning unit. This column is mandatory.

"boundary"

numeric "length" of shared boundary between the planning units identified on the same row. This column is mandatory.

Value

An object of class ConservationProblem-class().

Details

If you are familiar with using conservation planning software called Marxan, you should know that most of the arguments in this function follow the conventions used by the Marxan input files. This means that, pu corresponds to the data in the Planning Unit File (typically called "pu.dat"); features corresponds to the data in the Conservation Feature File (typically called "spec.dat"); dist_features corresponds to the data in the Planning Unit versus Conservation Feature File (typically called "dist_features.dat"); and boundary corresponds to the data in the Boundary Length File (typically called "bound.dat").

References

Ball I, Possingham H, Watts, M. Marxan and relatives: software for spatial conservation prioritization. Spatial conservation prioritisation: quantitative methods and computational tools 2009; 185<U+2012>195.

See also

Examples

## Create a data instance for the multi-action conservation planning problem
## using data.frames that have been loaded into R. This example uses input files included
## into package.

## Set prioriactions path
prioriactions_path <- system.file("extdata/input/", package = "prioriactions")

## Load in planning unit data
pu_data <- data.table::fread(paste0(prioriactions_path,"pu.dat"), data.table = FALSE)
head(pu_data)
#>   id cost status
#> 1  1    3      0
#> 2  2    4      0
#> 3  3    3      0
#> 4  4    3      0
#> 5  5    3      0
#> 6  6    3      0

## Load in feature data
features_data <- data.table::fread(paste0(prioriactions_path,"features.dat"), data.table = FALSE)
head(spec_data)
#> Error in head(spec_data) object 'spec_data' not found

## Load in planning unit vs feature data
dist_features_data <- data.table::fread(paste0(prioriactions_path,"dist_features.dat"), data.table = FALSE)
head(dist_features_data)
#>   pu species amount
#> 1  1       1      1
#> 2  1       7      1
#> 3  1       8      1
#> 4  1       9      1
#> 5  2       1      1
#> 6  2       2      1

## Load in the threats data
threats_data <- data.table::fread(paste0(prioriactions_path,"threats.dat"), data.table = FALSE)
head(threats_data)
#>   id    name blm_actions
#> 1  1 threat1           0
#> 2  2 threat2           0
#> 3  3 threat3           0
#> 4  4 threat4           0

## Load in the threats distribution data
dist_threats_data <- data.table::fread(paste0(prioriactions_path,"dist_threats.dat"), data.table = FALSE)
head(dist_threats_data)
#>   pu threats amount cost status
#> 1  1       2      1    3      0
#> 2  1       4      1    3      0
#> 3  2       2      1    4      0
#> 4  2       3      1    4      0
#> 5  2       4      1    4      0
#> 6  3       2      1    3      0

## Load in the sensitivity data
sensitivity_data <- data.table::fread(paste0(prioriactions_path,"sensibility.dat"), data.table = FALSE)
#> Error in data.table::fread(paste0(prioriactions_path, "sensibility.dat"),     data.table = FALSE) File '/tmp/Rtmp4F16ds/temp_libpath15fd079915e84/prioriactions/extdata/input/sensibility.dat' does not exist or is non-readable. getwd()=='/home/josesalgr/Desktop/Conservation/11. Prioriactions_repo/prioriactions/docs/reference'
head(sensitivity_data)
#> Error in head(sensitivity_data) object 'sensitivity_data' not found

## Load in the boundary data
boundary_data <- data.table::fread(paste0(prioriactions_path,"bound.dat"), data.table = FALSE)
head(boundary_data)
#>   id1 id2 boundary
#> 1   6   3        1
#> 2   6   7        1
#> 3  18  10        1
#> 4  18  11        1
#> 5  19  12        1
#> 6  19  18        1

## Create data instance
problem_data <- problem(
  pu = pu_data, features = features_data, dist_features = dist_features_data,
  dist_threats = dist_threats_data, threats = threats_data, sensitivity = sensitivity_data,
  boundary = boundary_data
)
#> Error in h(simpleError(msg, call)) error in evaluating the argument 'sensitivity' in selecting a method for function 'problem': object 'sensitivity_data' not found

## Instance summary
print(problem_data)
#> Error in print(problem_data) object 'problem_data' not found