solve

Solve the optimization model associated with the multi-action conservation planning problem. This function is used to internally solve the mathematical model associated with the problem in reference, specifying the type of solver and providing a set of parameters to control the optimization process and the user's machine settings.

This function requires the Rsymphony package, the Rglpk package or the gurobi package, which host the solvers inside. These packages are installed by default when the prioriactions package is installed. Of these three, only the Gurobi solver needs a license to operate .

solve(
  a,
  solver = "",
  gap_limit = 0,
  time_limit = .Machine$integer.max,
  solution_limit = FALSE,
  cores = 2,
  verbose = TRUE,
  name_output_file = "output",
  output_file = TRUE,
  ...
)

Arguments

a
Object of class OptimizationProblem-class() that specifies the custom-tailor optimization model created for the problem of prioritization of multiple conservation actions. This object must be created using the min_costs() function.

solver
A string label that indicates which solver to use to solve the model created for the problem of prioritization of multiple conservation actions. The following solvers are supported: "gurobi", and "symphony". The default argument is "gurobi" (requires a license).

gap_limit
A numeric value (between 0 and 1) that represents the gap to optimality, i.e., a relative number that cause the optimizer to terminate when the difference between the upper and lower objective function bounds is less than the gap times the upper bound. For example, a value of 0.01 will result in the optimizer stopping when the difference between the bounds is 1 percent of the upper bound. The default argument is 0.0.

time_limit
A numeric value (non negative) that indicates the time limit to run the optimizer (in seconds). The solver will return the current best solution when this time limit is exceeded. The default argument is the maximum time the user's machine can hold out.

solution_limit
A logical value that indicates if the solution process should be stopped after the first feasible solution is found (TRUE()), or if it should continue its normal course until the optimal solution is found (FALSE()). The default argument is FALSE.

cores
An integer value that indicates the number of parallel cores to use in the machine. The default argument is 1 (only one core in use).

verbose
A logical value that indicates if the solver information is displayed while solving the optimization model (TRUE()), or if it is not displayed (FALSE()). The default argument is TRUE.

Value

An object of class Solution-class().

Details

The solvers supported by the solve() function are described below.

Default solver

This solver uses the best software currently installed on the system.

Gurobi solver

Gurobi is a state-of-the-art commercial optimization software with an R package interface. It is by far the fastest of the solvers available in this package, however, it is also the only solver that is not freely available. That said, licenses are available to academics at no cost. The gurobi package is distributed with the Gurobi software suite. This solver uses the gurobi package to solve problems.

Symphony solver

SYMPHONY is an open-source integer programming solver that is part of the Computational Infrastructure for Operations Research (COIN-OR) project, an initiative to promote development of open-source tools for operations research (a field that includes linear programming). The Rsymphony package provides an interface to COIN-OR and is available on CRAN. This solver uses the Rsymphony package to solve problems.

See also

Examples

## Solve a conservation problem using the associated optimization model and some solver
## that the user has available. This example uses input files included into package.

## Load data
data(example_pu_data, example_features_data, example_rij_data, example_threats_data, example_sensitivity_data, example_bound_data)
#> Warning: data set ‘example_pu_data’ not found
#> Warning: data set ‘example_features_data’ not found
#> Warning: data set ‘example_rij_data’ not found
#> Warning: data set ‘example_threats_data’ not found
#> Warning: data set ‘example_sensitivity_data’ not found
#> Warning: data set ‘example_bound_data’ not found

## Create data instance
problem_data <- problem(
  pu = example_pu_data, features = example_features_data, rij = example_rij_data,
  threats = example_threats_data, sensitivity = example_sensibility_data,
  bound = example_bound_data
)
#> Error in h(simpleError(msg, call)) error in evaluating the argument 'pu' in selecting a method for function 'problem': object 'example_pu_data' not found

## Create optimization model
problem_model <- min_costs(x = problem_data, blm = 1, blm_actions = 1)
#> Error in min_costs(x = problem_data, blm = 1, blm_actions = 1) could not find function "min_costs"

## Solve the optimization model using a default solver
model_solution <- solve(a = problem_model, verbose = FALSE)
#> Error in inherits(a, "OptimizationProblem") object 'problem_model' not found
model_solution
#> Error in eval(expr, envir, enclos) object 'model_solution' not found

## Solve the optimization model using the Gurobi solver
## NOTE: The Gurobi solver must be previously installed and must have a valid license!
model_solution <- solve(a = problem_model, solver = "gurobi")
#> Error in inherits(a, "OptimizationProblem") object 'problem_model' not found
model_solution
#> Error in eval(expr, envir, enclos) object 'model_solution' not found

## Solve the optimization model using the Symphony solver
model_solution <- solve(a = problem_model, solver = "symphony")
#> Error in inherits(a, "OptimizationProblem") object 'problem_model' not found
model_solution
#> Error in eval(expr, envir, enclos) object 'model_solution' not found

## Solution summary
print(model_solution)
#> Error in print(model_solution) object 'model_solution' not found

## Report the solution to the problem you are working on:
## (i) total cost of the conservation plan;
solution_report01 <- model_solution$getObjectiveValue()
#> Error in eval(expr, envir, enclos) object 'model_solution' not found
print(solution_report01)
#> Error in print(solution_report01) object 'solution_report01' not found

## (ii) planning units suggested to be included (value 1) and not included (value 0)
## within the conservation plan;
solution_report02 <- model_solution$getSolutionUnits()
#> Error in eval(expr, envir, enclos) object 'model_solution' not found
head(solution_report02)
#> Error in head(solution_report02) object 'solution_report02' not found

## (iii) conservation actions that are suggested to be carried out (value 1)
## and not carried out (value 0) within the conservation plan.
solution_report03 <- model_solution$getSolutionActions()
#> Error in eval(expr, envir, enclos) object 'model_solution' not found
head(solution_report03)
#> Error in head(solution_report03) object 'solution_report03' not found