Evaluate multiple blm values

Return one solution per instance for different values of blm. Like prioriactions() function, it inherits all arguments from inputData(), problem() and solve().

evalBlm(values = c(), ...)

Arguments

values
numeric. Values of blm to verify. More than one value is needed.

...
arguments inherited from inputData(), problem() and solve() functions.

Value

An object of class portfolio.

Details

evalblm() creates and solves multiple instances, of the corresponding multi-actions planning problem, for different values of blm. Alternatively, this could be obtained by executing function prioriactions() or by steps the inputData(), problem() and solve() functions; using, in each run, different blm values. However, the evalblm() function has two advantages with respect to this manual approach: : 1) it is more efficient to create the models (this is because the model is created just once and, at each iteration, only the blm values are updated); and 2) the output is a portfolio object, which allows obtaining information about the group of solutions (including all get functions).

Examples

# \donttest{
# set seed for reproducibility
set.seed(14)

## Create model and solve
port <- evalBlm(pu = sim_pu_data, features = sim_features_data,
                dist_features = sim_dist_features_data,
                threats = sim_threats_data,
                dist_threats = sim_dist_threats_data,
                sensitivity = sim_sensitivity_data,
                boundary = sim_boundary_data,
                values = c(0.0, 0.01, 0.02, 0.03),
                model_type = "minimizeCosts",
                time_limit = 50,
                output_file = FALSE,
                cores = 2)
#> *********************************
#>  Iteration 1 of 4: Blm0
#> *********************************
#> Warning: The blm argument was set to 0, so the boundary data has no effect
#> Warning: Some blm_actions argument were set to 0, so the boundary data has no effect for these cases
#> Gurobi Optimizer version 9.1.2 build v9.1.2rc0 (linux64)
#> Thread count: 2 physical cores, 4 logical processors, using up to 2 threads
#> Optimize a model with 284 rows, 396 columns and 785 nonzeros
#> Model fingerprint: 0xff18a474
#> Variable types: 176 continuous, 220 integer (220 binary)
#> Coefficient statistics:
#>   Matrix range     [5e-01, 2e+00]
#>   Objective range  [1e+00, 1e+01]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [2e+01, 5e+01]
#> Found heuristic solution: objective 964.0000000
#> Found heuristic solution: objective 886.0000000
#> Presolve removed 251 rows and 294 columns
#> Presolve time: 0.01s
#> Presolved: 33 rows, 102 columns, 194 nonzeros
#> Variable types: 0 continuous, 102 integer (89 binary)
#> 
#> Root relaxation: objective 7.345000e+02, 39 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0  734.50000    0    4  886.00000  734.50000  17.1%     -    0s
#> H    0     0                     742.0000000  734.50000  1.01%     -    0s
#>      0     0  737.00000    0    2  742.00000  737.00000  0.67%     -    0s
#> *    0     0               0     738.0000000  738.00000  0.00%     -    0s
#> 
#> Explored 1 nodes (67 simplex iterations) in 0.05 seconds
#> Thread count was 2 (of 4 available processors)
#> 
#> Solution count 4: 738 742 886 964 
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective 7.380000000000e+02, best bound 7.380000000000e+02, gap 0.0000%
#> *********************************
#>  Iteration 2 of 4: Blm0.01
#> *********************************
#> Warning: Some blm_actions argument were set to 0, so the boundary data has no effect for these cases
#> Gurobi Optimizer version 9.1.2 build v9.1.2rc0 (linux64)
#> Thread count: 2 physical cores, 4 logical processors, using up to 2 threads
#> Optimize a model with 29984 rows, 10296 columns and 70085 nonzeros
#> Model fingerprint: 0x53d3645d
#> Variable types: 176 continuous, 10120 integer (10120 binary)
#> Coefficient statistics:
#>   Matrix range     [5e-01, 2e+00]
#>   Objective range  [1e-02, 2e+01]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 5e+01]
#> Found heuristic solution: objective 964.0000000
#> Presolve removed 190 rows and 181 columns
#> Presolve time: 0.17s
#> Presolved: 29794 rows, 10115 columns, 69709 nonzeros
#> Variable types: 0 continuous, 10115 integer (10110 binary)
#> 
#> Root relaxation: objective 8.257523e+02, 6527 iterations, 0.76 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0  825.75234    0  672  964.00000  825.75234  14.3%     -    1s
#> H    0     0                     828.3563400  825.75234  0.31%     -    1s
#> H    0     0                     828.2124400  825.75234  0.30%     -    1s
#> *    0     0               0     827.3912600  827.39126  0.00%     -    1s
#> 
#> Cutting planes:
#>   Cover: 4
#>   Flow cover: 1
#>   RLT: 1
#> 
#> Explored 1 nodes (6620 simplex iterations) in 1.40 seconds
#> Thread count was 2 (of 4 available processors)
#> 
#> Solution count 4: 827.391 828.212 828.356 964 
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective 8.273912600000e+02, best bound 8.273912600000e+02, gap 0.0000%
#> *********************************
#>  Iteration 3 of 4: Blm0.02
#> *********************************
#> Warning: Some blm_actions argument were set to 0, so the boundary data has no effect for these cases
#> Gurobi Optimizer version 9.1.2 build v9.1.2rc0 (linux64)
#> Thread count: 2 physical cores, 4 logical processors, using up to 2 threads
#> Optimize a model with 29984 rows, 10296 columns and 70085 nonzeros
#> Model fingerprint: 0x4633ec81
#> Variable types: 176 continuous, 10120 integer (10120 binary)
#> Coefficient statistics:
#>   Matrix range     [5e-01, 2e+00]
#>   Objective range  [1e-02, 1e+01]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 5e+01]
#> Found heuristic solution: objective 479.0000000
#> Presolve removed 190 rows and 181 columns
#> Presolve time: 0.16s
#> Presolved: 29794 rows, 10115 columns, 69709 nonzeros
#> Variable types: 0 continuous, 10115 integer (10110 binary)
#> 
#> Root relaxation: objective 3.770000e+02, 115 iterations, 0.02 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0  377.00000    0    1  479.00000  377.00000  21.3%     -    0s
#> H    0     0                     380.0000000  377.00000  0.79%     -    0s
#> H    0     0                     378.0000000  377.00000  0.26%     -    0s
#> 
#> Cutting planes:
#>   Cover: 1
#>   RLT: 1
#> 
#> Explored 1 nodes (115 simplex iterations) in 0.34 seconds
#> Thread count was 2 (of 4 available processors)
#> 
#> Solution count 3: 378 380 479 
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective 3.780000000000e+02, best bound 3.780000000000e+02, gap 0.0000%
#> *********************************
#>  Iteration 4 of 4: Blm0.03
#> *********************************
#> Warning: Some blm_actions argument were set to 0, so the boundary data has no effect for these cases
#> Gurobi Optimizer version 9.1.2 build v9.1.2rc0 (linux64)
#> Thread count: 2 physical cores, 4 logical processors, using up to 2 threads
#> Optimize a model with 29984 rows, 10296 columns and 70085 nonzeros
#> Model fingerprint: 0x87596595
#> Variable types: 176 continuous, 10120 integer (10120 binary)
#> Coefficient statistics:
#>   Matrix range     [5e-01, 2e+00]
#>   Objective range  [1e-03, 1e+01]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 5e+01]
#> Found heuristic solution: objective -6.0000000
#> Presolve removed 190 rows and 181 columns
#> Presolve time: 0.16s
#> Presolved: 29794 rows, 10115 columns, 69709 nonzeros
#> Variable types: 0 continuous, 10115 integer (10110 binary)
#> 
#> Root relaxation: objective -1.080000e+02, 11 iterations, 0.01 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 -108.00000    0    1   -6.00000 -108.00000  1700%     -    0s
#> H    0     0                    -105.0000000 -108.00000  2.86%     -    0s
#> H    0     0                    -107.0000000 -108.00000  0.93%     -    0s
#>      0     0     cutoff    0      -107.00000 -107.00000  0.00%     -    0s
#> 
#> Cutting planes:
#>   Gomory: 1
#>   Cover: 2
#>   RLT: 2
#> 
#> Explored 1 nodes (12 simplex iterations) in 0.33 seconds
#> Thread count was 2 (of 4 available processors)
#> 
#> Solution count 3: -107 -105 -6 
#> No other solutions better than -107
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective -1.070000000000e+02, best bound -1.070000000000e+02, gap 0.0000%

getConnectivityPenalty(port)
#>   solution_name     units threat_1 threat_2
#> 1          Blm0 10299.201 2732.466 1556.262
#> 2       Blm0.01  6639.126 2984.187 1115.137
#> 3       Blm0.02         0 3051.239  921.108
#> 4       Blm0.03         0 3051.239  929.962
# }