Evaluate multiple budget values

Provides multiple solutions for different values of budgets. This function assumes that the maximizeBenefits option is being used (note that the minimizeCosts option does not require setting a maximum budget). Like prioriactions() function, it inherits all arguments from problem(), model() and solve().

evalBudget(data = list(), values = c(), ...)

Arguments

data
list. Input data list for problem() function.

values
numeric. Values of budget to verify. More than one value is needed.

...
arguments inherited from problem(), maximizeBenefits(), and solve() functions.

Value

An object of class portfolio.

Details

evalBudget() creates and solves multiple instances, of the corresponding multi-actions planning problem, for different values of maximum budgets. Alternatively, this could be obtained by executing function prioriactions() or by steps the problem(), model() and solve() functions; using, in each run, different budgtes values. However, the evalBudget() function has two advantages with respect to this manual approach: : 1) it is more efficient to create the models (this is because the model is created just once and, at each iteration, only the budget values are updated); and 2) the output is a portfolio object, which allows obtaining information about the group of solutions (including all get functions).

Examples

# set seed for reproducibility
set.seed(14)

## Load data
inputs <- list(sim_pu_data, sim_features_data, sim_dist_features_data,
sim_threats_data, sim_dist_threats_data, sim_sensitivity_data,
sim_boundary_data)

## Create model and solve
port <- evalBudget(data = inputs, values = c(1, 10, 50, 100), time_limit = 50, output_file = FALSE)
#> *********************************
#>  Iteration 1 of 4: Budget1
#> *********************************
#> Warning: The blm argument was set to 0, so the boundary data has no effect
#> Warning: Some blm_actions argument were set to 0, so the boundary data has no effect for these cases
#> Gurobi Optimizer version 9.1.2 build v9.1.2rc0 (linux64)
#> Thread count: 2 physical cores, 4 logical processors, using up to 2 threads
#> Optimize a model with 277 rows, 589 columns and 883 nonzeros
#> Model fingerprint: 0xa2ead4ab
#> Model has 176 general constraints
#> Variable types: 352 continuous, 237 integer (237 binary)
#> Coefficient statistics:
#>   Matrix range     [5e-01, 1e+01]
#>   Objective range  [1e+00, 1e+00]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 1e+00]
#> Found heuristic solution: objective -0.0000000
#> Presolve removed 277 rows and 589 columns
#> Presolve time: 0.00s
#> Presolve: All rows and columns removed
#> 
#> Explored 0 nodes (0 simplex iterations) in 0.00 seconds
#> Thread count was 1 (of 4 available processors)
#> 
#> Solution count 1: -0 
#> No other solutions better than -0
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective -0.000000000000e+00, best bound -0.000000000000e+00, gap 0.0000%
#> *********************************
#>  Iteration 2 of 4: Budget10
#> *********************************
#> Gurobi Optimizer version 9.1.2 build v9.1.2rc0 (linux64)
#> Thread count: 2 physical cores, 4 logical processors, using up to 2 threads
#> Optimize a model with 277 rows, 589 columns and 883 nonzeros
#> Model fingerprint: 0xe554d2d7
#> Model has 176 general constraints
#> Variable types: 352 continuous, 237 integer (237 binary)
#> Coefficient statistics:
#>   Matrix range     [5e-01, 1e+01]
#>   Objective range  [1e+00, 1e+00]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+01, 1e+01]
#> Found heuristic solution: objective -0.0000000
#> Presolve added 0 rows and 136 columns
#> Presolve removed 8 rows and 0 columns
#> Presolve time: 0.05s
#> Presolved: 269 rows, 725 columns, 1439 nonzeros
#> Presolved model has 121 SOS constraint(s)
#> Variable types: 605 continuous, 120 integer (117 binary)
#> 
#> Root relaxation: objective 6.666667e+00, 235 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0    6.66667    0    3   -0.00000    6.66667      -     -    0s
#> H    0     0                       6.0000000    6.66667  11.1%     -    0s
#>      0     0    6.50000    0    3    6.00000    6.50000  8.33%     -    0s
#>      0     0    6.50000    0    3    6.00000    6.50000  8.33%     -    0s
#>      0     0     cutoff    0         6.00000    6.00000  0.00%     -    0s
#> 
#> Cutting planes:
#>   Gomory: 1
#>   Cover: 1
#> 
#> Explored 1 nodes (277 simplex iterations) in 0.06 seconds
#> Thread count was 2 (of 4 available processors)
#> 
#> Solution count 2: 6 -0 
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective 6.000000000000e+00, best bound 6.000000000000e+00, gap 0.0000%
#> *********************************
#>  Iteration 3 of 4: Budget50
#> *********************************
#> Gurobi Optimizer version 9.1.2 build v9.1.2rc0 (linux64)
#> Thread count: 2 physical cores, 4 logical processors, using up to 2 threads
#> Optimize a model with 277 rows, 589 columns and 883 nonzeros
#> Model fingerprint: 0xdce53123
#> Model has 176 general constraints
#> Variable types: 352 continuous, 237 integer (237 binary)
#> Coefficient statistics:
#>   Matrix range     [5e-01, 1e+01]
#>   Objective range  [1e+00, 1e+00]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [5e+01, 5e+01]
#> Found heuristic solution: objective -0.0000000
#> Presolve added 113 rows and 459 columns
#> Presolve time: 0.00s
#> Presolved: 390 rows, 1048 columns, 2082 nonzeros
#> Presolved model has 176 SOS constraint(s)
#> Variable types: 880 continuous, 168 integer (162 binary)
#> 
#> Root relaxation: objective 2.640000e+01, 309 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0   26.40000    0    3   -0.00000   26.40000      -     -    0s
#> H    0     0                      26.0000000   26.40000  1.54%     -    0s
#>      0     0   26.40000    0    3   26.00000   26.40000  1.54%     -    0s
#>      0     0   26.40000    0    3   26.00000   26.40000  1.54%     -    0s
#>      0     0     cutoff    0        26.00000   26.00000  0.00%     -    0s
#> 
#> Cutting planes:
#>   Cover: 1
#>   MIR: 1
#>   RLT: 1
#> 
#> Explored 1 nodes (369 simplex iterations) in 0.02 seconds
#> Thread count was 2 (of 4 available processors)
#> 
#> Solution count 2: 26 -0 
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective 2.600000000000e+01, best bound 2.600000000000e+01, gap 0.0000%
#> *********************************
#>  Iteration 4 of 4: Budget100
#> *********************************
#> Gurobi Optimizer version 9.1.2 build v9.1.2rc0 (linux64)
#> Thread count: 2 physical cores, 4 logical processors, using up to 2 threads
#> Optimize a model with 277 rows, 589 columns and 883 nonzeros
#> Model fingerprint: 0x219109a4
#> Model has 176 general constraints
#> Variable types: 352 continuous, 237 integer (237 binary)
#> Coefficient statistics:
#>   Matrix range     [5e-01, 1e+01]
#>   Objective range  [1e+00, 1e+00]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+02, 1e+02]
#> Found heuristic solution: objective -0.0000000
#> Presolve added 113 rows and 459 columns
#> Presolve time: 0.00s
#> Presolved: 390 rows, 1048 columns, 2082 nonzeros
#> Presolved model has 176 SOS constraint(s)
#> Variable types: 880 continuous, 168 integer (162 binary)
#> 
#> Root relaxation: objective 4.500000e+01, 370 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0   45.00000    0    4   -0.00000   45.00000      -     -    0s
#> H    0     0                      44.0000000   45.00000  2.27%     -    0s
#> *    0     0               0      45.0000000   45.00000  0.00%     -    0s
#> 
#> Cutting planes:
#>   Gomory: 1
#>   MIR: 1
#> 
#> Explored 1 nodes (372 simplex iterations) in 0.02 seconds
#> Thread count was 2 (of 4 available processors)
#> 
#> Solution count 3: 45 44 -0 
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective 4.500000000000e+01, best bound 4.500000000000e+01, gap 0.0000%