Create and solve multi-actions planning problems

Create and solve a multi-actions planning problem. It can be used instead of following the sequence of the inputData(), problem() and solve() functions.

prioriactions(...)

Arguments

...

arguments inherited from inputData(), problem() and solve() functions.

Value

An object of class solution.

Examples

# \donttest{
## This example uses input files included into package.

## set seed for reproducibility
set.seed(14)

## Load data
data(sim_pu_data, sim_features_data, sim_dist_features_data,
sim_threats_data, sim_dist_threats_data, sim_sensitivity_data,
sim_boundary_data)

## Create data instance
s <- prioriactions(pu = sim_pu_data, features = sim_features_data,
                dist_features = sim_dist_features_data,
                threats = sim_threats_data,
                dist_threats = sim_dist_threats_data,
                sensitivity = sim_sensitivity_data,
                boundary = sim_boundary_data,
                model_type = "minimizeCosts",
                time_limit = 50,
                output_file = FALSE,
                cores = 2)
#> Warning: The blm argument was set to 0, so the boundary data has no effect
#> Warning: Some blm_actions argument were set to 0, so the boundary data has no effect for these cases
#> Rcplex: num variables=396 num constraints=284

print(s)
#> Solution overview
#>   name: sol
#>   objective value: 738
#>   gap:  0%
#>   status:  Optimal solution (according to gap tolerance: 0)
#>   runtime: 0.013 sec
# }