solve

Solve the optimization model associated with the multi-action conservation planning problem. This function is used to internally solve the mathematical model associated with the problem in reference, specifying the type of solver and providing a set of parameters to control the optimization process and the user's machine settings.

This function requires the Rsymphony package, the Rglpk package or the gurobi package, which host the solvers inside. These packages are installed by default when the prioriactions package is installed. Of these three, only the Gurobi solver needs a license to operate .

# S4 method for OptimizationProblem,missing
solve(
  a,
  b,
  solver = "",
  gap_limit = 0,
  time_limit = .Machine$integer.max,
  solution_limit = FALSE,
  cores = 2,
  verbose = TRUE,
  txt_file = FALSE,
  ...
)

Arguments

a
Object of class OptimizationProblem-class that specifies the custom-tailor optimization model created for the problem of prioritization of multiple conservation actions. This object must be created using the min_costs function.

solver
A string label that indicates which solver to use to solve the model created for the problem of prioritization of multiple conservation actions. The following solvers are supported: "gurobi", and "symphony". The default argument is "gurobi" (requires a license).

gap_limit
A numeric value (between 0 and 1) that represents the gap to optimality, i.e., a relative number that cause the optimizer to terminate when the difference between the upper and lower objective function bounds is less than the gap times the upper bound. For example, a value of 0.01 will result in the optimizer stopping when the difference between the bounds is 1 percent of the upper bound. The default argument is 0.0.

time_limit
A numeric value (non negative) that indicates the time limit to run the optimizer (in seconds). The solver will return the current best solution when this time limit is exceeded. The default argument is the maximum time the user's machine can hold out.

solution_limit
A logical value that indicates if the solution process should be stopped after the first feasible solution is found (TRUE), or if it should continue its normal course until the optimal solution is found (FALSE). The default argument is FALSE.

cores
An integer value that indicates the number of parallel cores to use in the machine. The default argument is 1 (only one core in use).

verbose
A logical value that indicates if the solver information is displayed while solving the optimization model (TRUE), or if it is not displayed (FALSE). The default argument is TRUE.

Value

An object of class Solution-class.

Details

The solvers supported by the solve function are described below.

Default solver

This solver uses the best software currently installed on the system.

Gurobi solver

Gurobi is a state-of-the-art commercial optimization software with an R package interface. It is by far the fastest of the solvers available in this package, however, it is also the only solver that is not freely available. That said, licenses are available to academics at no cost. The gurobi package is distributed with the Gurobi software suite. This solver uses the gurobi package to solve problems.

Symphony solver

SYMPHONY is an open-source integer programming solver that is part of the Computational Infrastructure for Operations Research (COIN-OR) project, an initiative to promote development of open-source tools for operations research (a field that includes linear programming). The Rsymphony package provides an interface to COIN-OR and is available on CRAN. This solver uses the Rsymphony package to solve problems.

See also

Examples

## Solve a conservation problem using the associated optimization model and some solver
## that the user has available. This example uses input files included into package.

## Load data
data(example_pu_data, example_features_data, example_rij_data, example_threats_data, example_sensibility_data, example_bound_data)

## Create data instance
problem_data <- problem(
  pu = example_pu_data, features = example_features_data, rij = example_rij_data,
  threats = example_threats_data, sensibility = example_sensibility_data,
  bound = example_bound_data
)
#> Warning: The following pu's do not contain species: 59 97 109 118 121 122 123 127 129 134 136 139 141 142 146 166 167 185 194 195 196 198 211 230
#> Correctly loaded inputs

## Create optimization model
problem_model <- min_costs(x = problem_data, blm = 1, blm_actions = 1)

## Solve the optimization model using a default solver
model_solution <- solve(a = problem_model, verbose = FALSE)
model_solution
#> Solution overview
#>   objective value: 1755.42968110999
#>   gap:  0%
#>   status:  Optimal solution (according to gap tolerance: 0%)
#>   runtime: 0.6 sec

## Solve the optimization model using the Gurobi solver
## NOTE: The Gurobi solver must be previously installed and must have a valid license!
model_solution <- solve(a = problem_model, solver = "gurobi")
#> Gurobi Optimizer version 9.1.2 build v9.1.2rc0 (linux64)
#> Thread count: 2 physical cores, 4 logical processors, using up to 2 threads
#> Optimize a model with 18004 rows, 11740 columns and 51487 nonzeros
#> Model fingerprint: 0x591b64eb
#> Variable types: 4016 continuous, 7724 integer (7724 binary)
#> Coefficient statistics:
#>   Matrix range     [4e-02, 4e+00]
#>   Objective range  [6e-01, 1e+01]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 1e+02]
#> Presolve removed 15217 rows and 9879 columns
#> Presolve time: 0.10s
#> Presolved: 2787 rows, 1861 columns, 8338 nonzeros
#> Variable types: 528 continuous, 1333 integer (1333 binary)
#> Found heuristic solution: objective 1819.3548490
#> 
#> Root relaxation: objective 1.749987e+03, 401 iterations, 0.01 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 1749.98694    0  103 1819.35485 1749.98694  3.81%     -    0s
#> H    0     0                    1781.0872643 1749.98694  1.75%     -    0s
#>      0     0 1751.62933    0  122 1781.08726 1751.62933  1.65%     -    0s
#> H    0     0                    1779.7859349 1751.62933  1.58%     -    0s
#> H    0     0                    1765.5014779 1751.62933  0.79%     -    0s
#>      0     0 1752.46764    0  113 1765.50148 1752.46764  0.74%     -    0s
#>      0     0 1752.49813    0  113 1765.50148 1752.49813  0.74%     -    0s
#>      0     0 1752.80181    0  115 1765.50148 1752.80181  0.72%     -    0s
#> H    0     0                    1760.3790641 1752.80181  0.43%     -    0s
#> H    0     0                    1756.1364234 1752.80181  0.19%     -    0s
#>      0     0 1752.80181    0  115 1756.13642 1752.80181  0.19%     -    0s
#>      0     0 1752.93558    0  116 1756.13642 1752.93558  0.18%     -    0s
#>      0     0 1752.93558    0  123 1756.13642 1752.93558  0.18%     -    0s
#>      0     0 1753.37032    0  116 1756.13642 1753.37032  0.16%     -    0s
#>      0     0 1753.39048    0  109 1756.13642 1753.39048  0.16%     -    0s
#>      0     0 1753.39048    0  109 1756.13642 1753.39048  0.16%     -    0s
#>      0     0 1753.39327    0  116 1756.13642 1753.39327  0.16%     -    0s
#>      0     0 1753.43991    0  118 1756.13642 1753.43991  0.15%     -    0s
#>      0     0 1753.43991    0  118 1756.13642 1753.43991  0.15%     -    0s
#>      0     0 1753.46349    0  116 1756.13642 1753.46349  0.15%     -    0s
#>      0     0 1753.46349    0  120 1756.13642 1753.46349  0.15%     -    0s
#> H    0     0                    1755.4296811 1753.46349  0.11%     -    0s
#>      0     0 1753.46349    0  117 1755.42968 1753.46349  0.11%     -    0s
#>      0     0 1753.51387    0  116 1755.42968 1753.51387  0.11%     -    0s
#>      0     0 1753.51775    0  116 1755.42968 1753.51775  0.11%     -    0s
#>      0     0 1753.51878    0  117 1755.42968 1753.51878  0.11%     -    0s
#>      0     0 1753.51878    0  117 1755.42968 1753.51878  0.11%     -    0s
#>      0     2 1753.55948    0  117 1755.42968 1753.55948  0.11%     -    0s
#> 
#> Cutting planes:
#>   Gomory: 11
#>   Cover: 2
#>   MIR: 12
#>   RLT: 2
#> 
#> Explored 5 nodes (1581 simplex iterations) in 0.62 seconds
#> Thread count was 2 (of 4 available processors)
#> 
#> Solution count 7: 1755.43 1756.14 1760.38 ... 1819.35
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective 1.755429681110e+03, best bound 1.755429681110e+03, gap 0.0000%
model_solution
#> Solution overview
#>   objective value: 1755.42968110999
#>   gap:  0%
#>   status:  Optimal solution (according to gap tolerance: 0%)
#>   runtime: 0.62 sec

## Solve the optimization model using the Symphony solver
model_solution <- solve(a = problem_model, solver = "symphony")
model_solution
#> Solution overview
#>   objective value: 1755.42968111
#>   gap:  0%
#>   status:  Optimal solution (according to gap tolerance: 0%)
#>   runtime: 6.94 sec

## Solution summary
print(model_solution)
#> Solution overview
#>   objective value: 1755.42968111
#>   gap:  0%
#>   status:  Optimal solution (according to gap tolerance: 0%)
#>   runtime: 6.94 sec

## Report the solution to the problem you are working on:
## (i) total cost of the conservation plan;
solution_report01 <- model_solution$getObjetiveValue()
print(solution_report01)
#> [1] 1755.43

## (ii) planning units suggested to be included (value 1) and not included (value 0)
## within the conservation plan;
solution_report02 <- model_solution$getSolutionUnits()
head(solution_report02)
#>   id cost status solution
#> 1  1    3      0        0
#> 2  2    4      0        1
#> 3  3    3      0        1
#> 4  4    3      0        0
#> 5  5    3      0        1
#> 6  6    3      0        1

## (iii) conservation actions that are suggested to be carried out (value 1)
## and not carried out (value 0) within the conservation plan.
solution_report03 <- model_solution$getSolutionActions()
head(solution_report03)
#>   id solution_threat_1 solution_threat_2 solution_threat_3 solution_threat_4
#> 1  1                 0                 0                 0                 0
#> 2  2                 0                 1                 1                 1
#> 3  3                 0                 1                 0                 1
#> 4  4                 0                 0                 0                 0
#> 5  5                 0                 1                 0                 1
#> 6  6                 0                 1                 0                 1