Solve mathematical models

Solve the optimization model associated with the multi-action conservation planning problem. This function is used to solve the mathematical model created by the minimizeCosts() or maximizeBenefits() functions.

solve(
  a,
  solver = "",
  gap_limit = 0,
  time_limit = .Machine$integer.max,
  solution_limit = FALSE,
  cores = 2,
  verbose = TRUE,
  name_output_file = "output",
  output_file = TRUE
)

Arguments

a
optimizationProblem object. Optimization model created for the problem of prioritization of multiple conservation actions. This object must be created using the minimizeCosts() or maximizeBenefits() functions.

solver
string. Name of solver to use to solve the model. The following solvers are supported: "gurobi"(requires the gurobi package), and "symphony"(requires the Rsymphony package). We highly recommend using gurobi (for more information on how to obtain an academic license here).

gap_limit
numeric. Value between 0 and 1 that represents the gap to optimality, i.e., a relative number that cause the optimizer to terminate when the difference between the upper and lower objective function bounds is less than the gap times the upper bound. For example, a value of 0.01 will result in the optimizer stopping when the difference between the bounds is 1 percent of the upper bound.

time_limit
numeric. Time limit to run the optimizer (in seconds). The solver will return the current best solution when this time limit is exceeded.

solution_limit
logical. Indicates if the solution process should be stopped after the first feasible solution is found (TRUE), or not (FALSE).

cores
integer. Number of parallel cores to use in the machine to solve the problem.

verbose
logical. Indicates if the solver information is displayed while solving the optimization model (TRUE), or if it is not displayed (FALSE).

name_output_file
string. Prefix of all output names.

output_file
logical. Indicates if the outputs are exported as .csv files (TRUE), or they are not exported (FALSE). Currently, 5 files are exported. The distribution of actions in the solution, the distribution of the selected planning units, the benefits achieved by the features, the parameters used, and the optimization engine log.

Value

An object of class solution.

Details

The solvers supported by the solve() function are described below.

Gurobi solver

Gurobi is a state-of-the-art commercial optimization software with an R package interface. It is by far the fastest of the solvers available in this package, however, it is also the only solver that is not freely available. That said, licenses are available to academics at no cost. The gurobi package is distributed with the Gurobi software suite. This solver uses the gurobi package to solve problems.

Symphony solver

SYMPHONY is an open-source integer programming solver that is part of the Computational Infrastructure for Operations Research (COIN-OR) project, an initiative to promote development of open-source tools for operations research (a field that includes linear programming). The Rsymphony package provides an interface to COIN-OR and is available on CRAN. This solver uses the Rsymphony package to solve problems.

See also

Examples

## This example uses input files included into package.

## Load data
data(sim_pu_data, sim_features_data, sim_dist_features_data,
sim_threats_data, sim_dist_threats_data, sim_sensitivity_data,
sim_boundary_data)

## Create data instance
problem_data <- problem(
  pu = sim_pu_data, features = sim_features_data, dist_features = sim_dist_features_data,
  threats = sim_threats_data, dist_threats = sim_dist_threats_data, sensitivity = sim_sensitivity_data,
  boundary = sim_boundary_data
)

## Create optimization model
problem_model <- minimizeCosts(x = problem_data, blm = 1)
#> Warning: Some blm_actions argument were set to 0, so the boundary data has no effect for these cases

## Solve the optimization model using a gap_limit and gurobi solver
## NOTE: The Gurobi solver must be previously installed and must have a valid license!
s <- solve(a = problem_model, solver = "gurobi", gap_limit = 0.01, output_file = FALSE)
#> Gurobi Optimizer version 9.1.2 build v9.1.2rc0 (linux64)
#> Thread count: 2 physical cores, 4 logical processors, using up to 2 threads
#> Optimize a model with 29980 rows, 10489 columns and 70122 nonzeros
#> Model fingerprint: 0x3037b50b
#> Model has 176 general constraints
#> Variable types: 352 continuous, 10137 integer (10137 binary)
#> Coefficient statistics:
#>   Matrix range     [5e-01, 2e+00]
#>   Objective range  [1e+00, 1e+03]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 2e+01]
#> Found heuristic solution: objective 52870.722212
#> Presolve added 352 rows and 698 columns
#> Presolve time: 0.19s
#> Presolved: 30332 rows, 11187 columns, 72396 nonzeros
#> Presolved model has 176 SOS constraint(s)
#> Variable types: 1056 continuous, 10131 integer (10125 binary)
#> 
#> Root simplex log...
#> 
#> Iteration    Objective       Primal Inf.    Dual Inf.      Time
#>     5877    1.7317445e+04   1.895000e+02   0.000000e+00      1s
#>     7998    1.7570872e+04   0.000000e+00   0.000000e+00      1s
#> 
#> Root relaxation: objective 1.757087e+04, 7998 iterations, 1.13 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 17570.8718    0 4613 52870.7222 17570.8718  66.8%     -    1s
#> H    0     0                    18997.667675 17570.8718  7.51%     -    1s
#> H    0     0                    17810.201891 17570.8718  1.34%     -    1s
#>      0     0 17579.3062    0 4602 17810.2019 17579.3062  1.30%     -    2s
#>      0     0 17579.3062    0 2083 17810.2019 17579.3062  1.30%     -    3s
#>      0     0 17596.7487    0 2038 17810.2019 17596.7487  1.20%     -    4s
#>      0     0 17597.1098    0 2040 17810.2019 17597.1098  1.20%     -    4s
#>      0     0 17631.6491    0 2013 17810.2019 17631.6491  1.00%     -    4s
#>      0     0 17662.6335    0 1619 17810.2019 17662.6335  0.83%     -    5s
#> 
#> Cutting planes:
#>   Gomory: 3
#>   Cover: 9
#>   Implied bound: 9
#>   Flow cover: 8
#>   GUB cover: 1
#>   Mod-K: 4
#>   RLT: 1463
#>   BQP: 551
#> 
#> Explored 1 nodes (16454 simplex iterations) in 5.57 seconds
#> Thread count was 2 (of 4 available processors)
#> 
#> Solution count 3: 17810.2 18997.7 52870.7 
#> 
#> Optimal solution found (tolerance 1.00e-02)
#> Best objective 1.781020189131e+04, best bound 1.766263345760e+04, gap 0.8286%

print(s)
#> Solution overview
#>   name: sol
#>   objective value: 17810.2018913131
#>   gap:  0.8286
#>   status:  Optimal solution (according to gap tolerance: 0.8286)
#>   runtime: 5.57 sec

## Solve the optimization model using a time_limit and gurobi solver
s <- solve(a = problem_model, solver = "gurobi", time_limit = 10, output_file = FALSE)
#> Gurobi Optimizer version 9.1.2 build v9.1.2rc0 (linux64)
#> Thread count: 2 physical cores, 4 logical processors, using up to 2 threads
#> Optimize a model with 29980 rows, 10489 columns and 70122 nonzeros
#> Model fingerprint: 0x3037b50b
#> Model has 176 general constraints
#> Variable types: 352 continuous, 10137 integer (10137 binary)
#> Coefficient statistics:
#>   Matrix range     [5e-01, 2e+00]
#>   Objective range  [1e+00, 1e+03]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 2e+01]
#> Found heuristic solution: objective 52870.722212
#> Presolve added 352 rows and 698 columns
#> Presolve time: 0.16s
#> Presolved: 30332 rows, 11187 columns, 72396 nonzeros
#> Presolved model has 176 SOS constraint(s)
#> Variable types: 1056 continuous, 10131 integer (10125 binary)
#> 
#> Root simplex log...
#> 
#> Iteration    Objective       Primal Inf.    Dual Inf.      Time
#>     6501    1.7433507e+04   3.785000e+02   0.000000e+00      1s
#>     7998    1.7570872e+04   0.000000e+00   0.000000e+00      1s
#> 
#> Root relaxation: objective 1.757087e+04, 7998 iterations, 1.09 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 17570.8718    0 4613 52870.7222 17570.8718  66.8%     -    1s
#> H    0     0                    18997.667675 17570.8718  7.51%     -    1s
#> H    0     0                    17810.201891 17570.8718  1.34%     -    1s
#>      0     0 17579.3062    0 4602 17810.2019 17579.3062  1.30%     -    2s
#>      0     0 17579.3062    0 2083 17810.2019 17579.3062  1.30%     -    3s
#>      0     0 17596.7487    0 2038 17810.2019 17596.7487  1.20%     -    4s
#>      0     0 17597.1098    0 2040 17810.2019 17597.1098  1.20%     -    4s
#>      0     0 17631.6491    0 2013 17810.2019 17631.6491  1.00%     -    4s
#>      0     0 17662.6335    0 1619 17810.2019 17662.6335  0.83%     -    5s
#>      0     0 17662.6335    0 1870 17810.2019 17662.6335  0.83%     -    6s
#>      0     0 17662.6335    0 1779 17810.2019 17662.6335  0.83%     -    6s
#>      0     0 17662.6335    0 1506 17810.2019 17662.6335  0.83%     -    7s
#>      0     0 17664.9903    0 1517 17810.2019 17664.9903  0.82%     -    7s
#>      0     0 17666.2991    0 1374 17810.2019 17666.2991  0.81%     -    7s
#>      0     0 17666.4953    0 1374 17810.2019 17666.4953  0.81%     -    7s
#>      0     0 17666.5714    0 1375 17810.2019 17666.5714  0.81%     -    7s
#>      0     0 17679.0256    0 1759 17810.2019 17679.0256  0.74%     -    8s
#>      0     0 17679.0256    0 1870 17810.2019 17679.0256  0.74%     -    8s
#>      0     0 17679.0256    0 1758 17810.2019 17679.0256  0.74%     -    8s
#>      0     0 17679.0256    0 1549 17810.2019 17679.0256  0.74%     -    9s
#>      0     0 17679.0256    0 1674 17810.2019 17679.0256  0.74%     -    9s
#>      0     0 17680.8114    0 1686 17810.2019 17680.8114  0.73%     -    9s
#>      0     0 17680.8384    0 1692 17810.2019 17680.8384  0.73%     -    9s
#>      0     0 17687.5001    0 1388 17810.2019 17687.5001  0.69%     -    9s
#>      0     0 17687.9356    0 1390 17810.2019 17687.9356  0.69%     -    9s
#> 
#> Cutting planes:
#>   Gomory: 8
#>   GUB cover: 1
#>   Zero half: 1
#>   Mod-K: 2
#>   RLT: 946
#>   BQP: 280
#> 
#> Explored 1 nodes (30108 simplex iterations) in 10.00 seconds
#> Thread count was 2 (of 4 available processors)
#> 
#> Solution count 3: 17810.2 18997.7 52870.7 
#> 
#> Time limit reached
#> Best objective 1.781020189131e+04, best bound 1.768793556378e+04, gap 0.6865%

print(s)
#> Solution overview
#>   name: sol
#>   objective value: 17810.2018913131
#>   gap:  0.6865
#>   status:  Feasible solution (according to time limit: 10 sec)
#>   runtime: 10 sec